% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APPENC01}
\alias{APPENC01}
\title{APPENC01}
\format{\preformatted{'data.frame':	113 obs. of  12 variables:
 $ V1 : int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2 : num  7.13 8.82 8.34 8.95 11.2 ...
 $ V3 : num  55.7 58.2 56.9 53.7 56.5 50.9 57.8 45.7 48.2 56.3 ...
 $ V4 : num  4.1 1.6 2.7 5.6 5.7 5.1 4.6 5.4 4.3 6.3 ...
 $ V5 : num  9 3.8 8.1 18.9 34.5 21.9 16.7 60.5 24.4 29.6 ...
 $ V6 : num  39.6 51.7 74 122.8 88.9 ...
 $ V7 : int  279 80 107 147 180 150 186 640 182 85 ...
 $ V8 : int  2 2 2 2 2 2 2 1 2 2 ...
 $ V9 : int  4 2 3 4 1 2 3 2 3 1 ...
 $ V10: int  207 51 82 53 134 147 151 399 130 59 ...
 $ V11: int  241 52 54 148 151 106 129 360 118 66 ...
 $ V12: num  60 40 20 40 40 40 40 60 40 40 ...
}}
\usage{
APPENC01
}
\description{
APPENC01
}
\keyword{datasets}

