% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APPENC02}
\alias{APPENC02}
\title{APPENC02}
\format{\preformatted{'data.frame':	440 obs. of  17 variables:
 $ V1 : int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2 : Factor w/ 373 levels "Ada","Adams",..: 179 77 130 284 237 159 190 353 80 82 ...
 $ V3 : Factor w/ 48 levels "AL","AR","AZ",..: 4 13 42 4 4 33 3 21 9 42 ...
 $ V4 : int  4060 946 1729 4205 790 71 9204 614 1945 880 ...
 $ V5 : int  8863164 5105067 2818199 2498016 2410556 2300664 2122101 2111687 1937094 1852810 ...
 $ V6 : num  32.1 29.2 31.3 33.5 32.6 28.3 29.2 27.4 27.1 32.6 ...
 $ V7 : num  9.7 12.4 7.1 10.9 9.2 12.4 12.5 12.5 13.9 8.2 ...
 $ V8 : int  23677 15153 7553 5905 6062 4861 4320 3823 6274 4718 ...
 $ V9 : int  27700 21550 12449 6179 6369 8942 6104 9490 8840 6934 ...
 $ V10: int  688936 436936 253526 173821 144524 680966 177593 193978 244725 214258 ...
 $ V11: num  70 73.4 74.9 81.9 81.2 63.7 81.5 70 65 77.1 ...
 $ V12: num  22.3 22.8 25.4 25.3 27.8 16.6 22.1 13.7 18.8 26.3 ...
 $ V13: num  11.6 11.1 12.5 8.1 5.2 19.5 8.8 16.9 14.2 10.4 ...
 $ V14: num  8 7.2 5.7 6.1 4.8 9.5 4.9 10 8.7 6.1 ...
 $ V15: int  20786 21729 19517 19588 24400 16803 18042 17461 17823 21001 ...
 $ V16: int  184230 110928 55003 48931 58818 38658 38287 36872 34525 38911 ...
 $ V17: int  4 2 3 4 4 1 4 2 3 3 ...
}}
\usage{
APPENC02
}
\description{
APPENC02
}
\keyword{datasets}

