% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APPENC03}
\alias{APPENC03}
\title{APPENC03}
\format{\preformatted{'data.frame':	36 obs. of  8 variables:
 $ V1: int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2: num  3.15 2.52 2.64 2.55 2.69 2.38 3.02 2.52 2.45 2.42 ...
 $ V3: num  2.2 2.19 2.29 2.42 2.18 ...
 $ V4: int  498 510 422 858 566 536 585 310 211 504 ...
 $ V5: int  1 0 1 0 1 0 1 1 0 0 ...
 $ V6: int  1 0 1 1 0 0 1 0 0 1 ...
 $ V7: Factor w/ 12 levels "Apr","Aug","Dec",..: 12 11 10 3 5 4 8 1 9 7 ...
 $ V8: int  1999 1999 1999 1999 2000 2000 2000 2000 2000 2000 ...
}}
\usage{
APPENC03
}
\description{
APPENC03
}
\keyword{datasets}

