% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APPENC04}
\alias{APPENC04}
\title{APPENC04}
\format{\preformatted{'data.frame':	705 obs. of  5 variables:
 $ V1: int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2: num  0.98 1.13 1.25 1.32 1.48 1.57 1.67 1.73 1.75 1.76 ...
 $ V3: int  61 84 74 95 77 47 67 69 83 72 ...
 $ V4: int  20 20 19 23 28 23 28 20 22 24 ...
 $ V5: int  1996 1996 1996 1996 1996 1996 1996 1996 1996 1996 ...
}}
\usage{
APPENC04
}
\description{
APPENC04
}
\keyword{datasets}

