% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APPENC05}
\alias{APPENC05}
\title{APPENC05}
\format{\preformatted{'data.frame':	97 obs. of  9 variables:
 $ V1: int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2: num  0.651 0.852 0.852 0.852 1.448 ...
 $ V3: num  0.56 0.372 0.601 0.301 2.117 ...
 $ V4: num  16 27.7 14.7 26.6 30.9 ...
 $ V5: int  50 58 74 58 62 50 64 58 47 63 ...
 $ V6: num  0 0 0 0 0 ...
 $ V7: int  0 0 0 0 0 0 0 0 0 0 ...
 $ V8: num  0 0 0 0 0 0 0 0 0 0 ...
 $ V9: int  6 7 7 6 6 6 6 6 7 6 ...
}}
\usage{
APPENC05
}
\description{
APPENC05
}
\keyword{datasets}

