% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APPENC06}
\alias{APPENC06}
\title{APPENC06}
\format{\preformatted{'data.frame':	73 obs. of  8 variables:
 $ V1: int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2: int  1 2 7 2 1 10 10 1 1 6 ...
 $ V3: int  12 18 26 28 36 45 36 18 25 28 ...
 $ V4: int  1 1 1 1 1 1 1 2 2 2 ...
 $ V5: int  3 6 9 12 15 18 21 3 6 9 ...
 $ V6: int  0 0 0 0 0 1 1 0 0 0 ...
 $ V7: int  2001 2001 2001 2001 2002 2002 2002 2001 2001 2001 ...
 $ V8: int  1 2 3 4 1 2 3 1 2 3 ...
}}
\usage{
APPENC06
}
\description{
APPENC06
}
\keyword{datasets}

