% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APPENC07}
\alias{APPENC07}
\title{APPENC07}
\format{\preformatted{'data.frame':	522 obs. of  13 variables:
 $ V1 : int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2 : int  360000 340000 250000 205500 275500 248000 229900 150000 195000 160000 ...
 $ V3 : int  3032 2058 1780 1638 2196 1966 2216 1597 1622 1976 ...
 $ V4 : int  4 4 4 4 4 4 3 2 3 3 ...
 $ V5 : int  4 2 3 2 3 3 2 1 2 3 ...
 $ V6 : int  1 1 1 1 1 1 1 1 1 0 ...
 $ V7 : int  2 2 2 2 2 5 2 1 2 1 ...
 $ V8 : int  0 0 0 0 0 1 0 0 0 0 ...
 $ V9 : int  1972 1976 1980 1963 1968 1972 1972 1955 1975 1918 ...
 $ V10: int  2 2 2 2 2 2 2 2 3 3 ...
 $ V11: int  1 1 1 1 7 1 7 1 1 1 ...
 $ V12: int  22221 22912 21345 17342 21786 18902 18639 22112 14321 32358 ...
 $ V13: int  0 0 0 0 0 0 0 0 0 0 ...
}}
\usage{
APPENC07
}
\description{
APPENC07
}
\keyword{datasets}

