% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APPENC08}
\alias{APPENC08}
\title{APPENC08}
\format{\preformatted{'data.frame':	43 obs. of  10 variables:
 $ V1 : int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2 : int  121 227 446 731 1017 1473 933 2004 3224 1176 ...
 $ V3 : num  0.075 0.075 0.075 0.075 0.075 0.075 0.08 0.08 0.0825 0.0825 ...
 $ V4 : int  64 64 65 65 65 68 69 68 70 68 ...
 $ V5 : int  0 0 0 0 0 0 0 0 3 0 ...
 $ V6 : int  3536 3042 2456 2198 1835 2027 2360 2270 3106 3942 ...
 $ V7 : int  615 813 704 1094 825 1140 1023 1168 2388 1024 ...
 $ V8 : num  2.22 0.28 0.79 0.79 3.42 2.14 2.33 0.02 0.42 0.19 ...
 $ V9 : int  1999 1999 1999 1999 1999 1999 1999 1999 1999 1999 ...
 $ V10: int  1 2 3 4 5 6 7 8 9 10 ...
}}
\usage{
APPENC08
}
\description{
APPENC08
}
\keyword{datasets}

