% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APPENC09}
\alias{APPENC09}
\title{APPENC09}
\format{\preformatted{'data.frame':	788 obs. of  10 variables:
 $ V1 : int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2 : num  179 319 9311 281 18727 ...
 $ V3 : int  63 59 62 60 55 66 64 45 68 64 ...
 $ V4 : int  0 0 0 1 0 0 1 1 0 1 ...
 $ V5 : int  2 2 17 9 5 1 2 3 6 3 ...
 $ V6 : int  1 0 0 0 2 0 0 0 2 0 ...
 $ V7 : int  4 6 2 7 7 3 3 5 5 2 ...
 $ V8 : int  0 0 0 0 0 0 0 0 0 0 ...
 $ V9 : int  3 0 5 2 0 4 1 1 4 0 ...
 $ V10: int  300 120 353 332 18 296 247 82 334 85 ...
}}
\usage{
APPENC09
}
\description{
APPENC09
}
\keyword{datasets}

