% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APPENC10}
\alias{APPENC10}
\title{APPENC10}
\format{\preformatted{'data.frame':	196 obs. of  6 variables:
 $ V1: int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2: int  33 35 6 60 18 26 6 31 26 37 ...
 $ V3: int  1 1 1 1 3 3 3 2 2 2 ...
 $ V4: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V5: int  0 0 0 0 1 0 0 1 1 0 ...
 $ V6: int  1 1 0 1 0 0 0 1 0 0 ...
}}
\usage{
APPENC10
}
\description{
APPENC10
}
\keyword{datasets}

