% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APPENC11}
\alias{APPENC11}
\title{APPENC11}
\format{\preformatted{'data.frame':	482 obs. of  5 variables:
 $ V1: int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2: int  0 0 0 0 0 0 0 1 0 0 ...
 $ V3: int  1200000 1454000 1500000 1530000 2000000 2100000 2100000 2500000 2500000 2750000 ...
 $ V4: int  3000000 1454000 300000 510000 800000 700000 400000 500000 500000 550000 ...
 $ V5: int  0 1 0 0 0 0 0 0 0 0 ...
}}
\usage{
APPENC11
}
\description{
APPENC11
}
\keyword{datasets}

