% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APPENC12}
\alias{APPENC12}
\title{APPENC12}
\format{\preformatted{'data.frame':	192 obs. of  7 variables:
 $ V1: int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V4: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V5: int  1 2 3 4 1 2 3 4 1 2 ...
 $ V6: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V7: num  0.81 0.8 0.82 0.5 0.77 0.78 0.79 0.51 0.8 0.82 ...
}}
\usage{
APPENC12
}
\description{
APPENC12
}
\keyword{datasets}

