% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APPEND10}
\alias{APPEND10}
\title{APPEND10}
\format{\preformatted{'data.frame':	27 obs. of  4 variables:
 $ V1: num  28 32 31 26 24 27 27 29 30 24 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 2 ...
 $ V3: int  1 1 1 2 2 2 3 3 3 1 ...
 $ V4: int  1 2 3 1 2 3 1 2 3 1 ...
}}
\usage{
APPEND10
}
\description{
APPEND10
}
\keyword{datasets}

