% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH01PR19}
\alias{CH01PR19}
\title{CH01PR19}
\format{\preformatted{'data.frame':	120 obs. of  2 variables:
 $ V1: num  3.9 3.88 3.78 2.54 3.03 ...
 $ V2: int  21 14 28 22 21 31 32 27 29 26 ...
}}
\usage{
CH01PR19
}
\description{
CH01PR19
}
\keyword{datasets}

