% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH01PR20}
\alias{CH01PR20}
\title{CH01PR20}
\format{\preformatted{'data.frame':	45 obs. of  2 variables:
 $ V1: int  20 60 46 41 12 137 68 89 4 32 ...
 $ V2: int  2 4 3 2 1 10 5 5 1 2 ...
}}
\usage{
CH01PR20
}
\description{
CH01PR20
}
\keyword{datasets}

