% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH01PR21}
\alias{CH01PR21}
\title{CH01PR21}
\format{\preformatted{'data.frame':	10 obs. of  2 variables:
 $ V1: num  16 9 17 12 22 13 8 15 19 11
 $ V2: num  1 0 2 0 3 1 0 1 2 0
}}
\usage{
CH01PR21
}
\description{
CH01PR21
}
\keyword{datasets}

