% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH01PR22}
\alias{CH01PR22}
\title{CH01PR22}
\format{\preformatted{'data.frame':	16 obs. of  2 variables:
 $ V1: num  199 205 196 200 218 220 215 223 237 234 ...
 $ V2: num  16 16 16 16 24 24 24 24 32 32 ...
}}
\usage{
CH01PR22
}
\description{
CH01PR22
}
\keyword{datasets}

