% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH01PR27}
\alias{CH01PR27}
\title{CH01PR27}
\format{\preformatted{'data.frame':	60 obs. of  2 variables:
 $ V1: int  106 106 97 113 96 119 92 112 92 102 ...
 $ V2: int  43 41 47 46 45 41 47 41 48 48 ...
}}
\usage{
CH01PR27
}
\description{
CH01PR27
}
\keyword{datasets}

