% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH01PR28}
\alias{CH01PR28}
\title{CH01PR28}
\format{\preformatted{'data.frame':	84 obs. of  2 variables:
 $ V1: int  8487 8179 8362 8220 6246 9100 6561 5873 7993 7932 ...
 $ V2: int  74 82 81 81 87 66 68 81 74 82 ...
}}
\usage{
CH01PR28
}
\description{
CH01PR28
}
\keyword{datasets}

