% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH01PR42}
\alias{CH01PR42}
\title{CH01PR42}
\format{\preformatted{'data.frame':	6 obs. of  2 variables:
 $ V1: num  128 213 75 250 446 540
 $ V2: num  7 12 4 14 25 30
}}
\usage{
CH01PR42
}
\description{
CH01PR42
}
\keyword{datasets}

