% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH02PR42}
\alias{CH02PR42}
\title{CH02PR42}
\format{\preformatted{'data.frame':	15 obs. of  2 variables:
 $ V1: num  13.9 16 10.3 11.8 16.7 12.5 10 11.4 13.9 12.2 ...
 $ V2: num  28.6 34.7 21 25.5 36.8 24 19.1 22.5 28.3 25 ...
}}
\usage{
CH02PR42
}
\description{
CH02PR42
}
\keyword{datasets}

