% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH02TA04}
\alias{CH02TA04}
\title{CH02TA04}
\format{\preformatted{'data.frame':	12 obs. of  2 variables:
 $ V1: int  29 435 86 1090 219 503 47 3524 185 98 ...
 $ V2: int  127 214 133 208 153 184 130 217 141 154 ...
}}
\usage{
CH02TA04
}
\description{
CH02TA04
}
\keyword{datasets}

