% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH03PR03}
\alias{CH03PR03}
\title{CH03PR03}
\format{\preformatted{'data.frame':	120 obs. of  4 variables:
 $ V1: num  3.9 3.88 3.78 2.54 3.03 ...
 $ V2: int  21 14 28 22 21 31 32 27 29 26 ...
 $ V3: int  122 132 119 99 131 139 113 136 75 106 ...
 $ V4: int  99 71 95 75 46 77 85 99 13 97 ...
}}
\usage{
CH03PR03
}
\description{
CH03PR03
}
\keyword{datasets}

