% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH03PR04}
\alias{CH03PR04}
\title{CH03PR04}
\format{\preformatted{'data.frame':	45 obs. of  4 variables:
 $ V1: int  20 60 46 41 12 137 68 89 4 32 ...
 $ V2: int  2 4 3 2 1 10 5 5 1 2 ...
 $ V3: int  20 19 27 32 24 26 24 28 22 18 ...
 $ V4: int  4 5 4 1 4 4 3 6 6 6 ...
}}
\usage{
CH03PR04
}
\description{
CH03PR04
}
\keyword{datasets}

