% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH03PR09}
\alias{CH03PR09}
\title{CH03PR09}
\format{\preformatted{'data.frame':	10 obs. of  2 variables:
 $ V1: int  2 3 4 5 6 7 8 9 10 11
 $ V2: num  3.2 2.9 -1.7 -2 -2.3 -1.2 -0.9 0.8 0.7 0.5
}}
\usage{
CH03PR09
}
\description{
CH03PR09
}
\keyword{datasets}

