% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH03PR10}
\alias{CH03PR10}
\title{CH03PR10}
\format{\preformatted{'data.frame':	12 obs. of  2 variables:
 $ V1: num  9.9 9.3 10.2 9.6 10.2 12.4 14.3 9.6 9.2 15.6 ...
 $ V2: num  -1.12 0.81 -0.76 0.43 0.65 -0.17 1.62 1.79 -0.53 -3.78 ...
}}
\usage{
CH03PR10
}
\description{
CH03PR10
}
\keyword{datasets}

