% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH03PR11}
\alias{CH03PR11}
\title{CH03PR11}
\format{\preformatted{'data.frame':	9 obs. of  2 variables:
 $ V1: int  -1 0 1 -1 0 1 -1 0 1
 $ V2: num  0.5 2.1 -3.4 0.3 -1.7 4.2 -0.6 2.6 -4
}}
\usage{
CH03PR11
}
\description{
CH03PR11
}
\keyword{datasets}

