% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH03PR15}
\alias{CH03PR15}
\title{CH03PR15}
\format{\preformatted{'data.frame':	15 obs. of  2 variables:
 $ V1: num  0.07 0.09 0.08 0.16 0.17 0.21 0.49 0.58 0.53 1.22 ...
 $ V2: num  9 9 9 7 7 7 5 5 5 3 ...
}}
\usage{
CH03PR15
}
\description{
CH03PR15
}
\keyword{datasets}

