% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH03PR17}
\alias{CH03PR17}
\title{CH03PR17}
\format{\preformatted{'data.frame':	10 obs. of  2 variables:
 $ V1: num  98 135 162 178 221 232 283 300 374 395
 $ V2: num  0 1 2 3 4 5 6 7 8 9
}}
\usage{
CH03PR17
}
\description{
CH03PR17
}
\keyword{datasets}

