% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH03PR18}
\alias{CH03PR18}
\title{CH03PR18}
\format{\preformatted{'data.frame':	111 obs. of  2 variables:
 $ V1: num  14.28 8.8 12.49 9.38 10.89 ...
 $ V2: int  15 9 7 4 9 21 11 6 10 7 ...
}}
\usage{
CH03PR18
}
\description{
CH03PR18
}
\keyword{datasets}

