% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH03PR24}
\alias{CH03PR24}
\title{CH03PR24}
\format{\preformatted{'data.frame':	8 obs. of  2 variables:
 $ V1: num  63 67 74 64 75 69 90 60
 $ V2: num  5 8 11 7 13 12 12 6
}}
\usage{
CH03PR24
}
\description{
CH03PR24
}
\keyword{datasets}

