% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH03TA01}
\alias{CH03TA01}
\title{CH03TA01}
\format{\preformatted{'data.frame':	8 obs. of  2 variables:
 $ V1: num  0.6 6.7 5.3 4 6.55 2.15 6.6 5.75
 $ V2: int  80 220 140 120 180 100 200 160
}}
\usage{
CH03TA01
}
\description{
CH03TA01
}
\keyword{datasets}

