% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH03TA04}
\alias{CH03TA04}
\title{CH03TA04}
\format{\preformatted{'data.frame':	11 obs. of  2 variables:
 $ V1: int  125 100 200 75 150 175 75 175 125 200 ...
 $ V2: int  160 112 124 28 152 156 42 124 150 104 ...
}}
\usage{
CH03TA04
}
\description{
CH03TA04
}
\keyword{datasets}

