% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH03TA07}
\alias{CH03TA07}
\title{CH03TA07}
\format{\preformatted{'data.frame':	10 obs. of  2 variables:
 $ V1: num  0.5 0.5 1 1 1.5 1.5 2 2 2.5 2.5
 $ V2: num  42.5 50.6 68.5 80.7 89 ...
}}
\usage{
CH03TA07
}
\description{
CH03TA07
}
\keyword{datasets}

