% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH03TA08}
\alias{CH03TA08}
\title{CH03TA08}
\format{\preformatted{'data.frame':	25 obs. of  3 variables:
 $ V1: num  0 0 0 0 0 1 1 1 1 1 ...
 $ V2: num  13.4 12.8 11.9 20.1 15.6 ...
 $ V3: num  1.13 1.11 1.08 1.3 1.19 ...
}}
\usage{
CH03TA08
}
\description{
CH03TA08
}
\keyword{datasets}

