% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH03TA10}
\alias{CH03TA10}
\title{CH03TA10}
\format{\preformatted{'data.frame':	24 obs. of  2 variables:
 $ V1: num  0.15 0.004 0.069 0.03 0.011 0.004 0.041 0.109 0.068 0.009 ...
 $ V2: int  20 0 10 5 0 0 5 20 10 0 ...
}}
\usage{
CH03TA10
}
\description{
CH03TA10
}
\keyword{datasets}

