% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH04PR12}
\alias{CH04PR12}
\title{CH04PR12}
\format{\preformatted{'data.frame':	12 obs. of  2 variables:
 $ V1: num  128 213 191 178 250 446 540 457 324 177 ...
 $ V2: num  7 12 10 10 14 25 30 25 18 10 ...
}}
\usage{
CH04PR12
}
\description{
CH04PR12
}
\keyword{datasets}

