% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH04TA02}
\alias{CH04TA02}
\title{CH04TA02}
\format{\preformatted{'data.frame':	12 obs. of  2 variables:
 $ V1: int  20 196 115 50 122 100 33 154 80 147 ...
 $ V2: int  114 921 560 245 575 475 138 727 375 670 ...
}}
\usage{
CH04TA02
}
\description{
CH04TA02
}
\keyword{datasets}

