% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH05PR04}
\alias{CH05PR04}
\title{CH05PR04}
\format{\preformatted{'data.frame':	5 obs. of  2 variables:
 $ V1: num  7.8 9 10.2 11 11.7
 $ V2: num  8 4 0 -4 -8
}}
\usage{
CH05PR04
}
\description{
CH05PR04
}
\keyword{datasets}

