% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH05PR05}
\alias{CH05PR05}
\title{CH05PR05}
\format{\preformatted{'data.frame':	6 obs. of  2 variables:
 $ V1: num  16 5 10 15 13 22
 $ V2: num  4 1 2 3 3 4
}}
\usage{
CH05PR05
}
\description{
CH05PR05
}
\keyword{datasets}

