% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH06PR05}
\alias{CH06PR05}
\title{CH06PR05}
\format{\preformatted{'data.frame':	16 obs. of  3 variables:
 $ V1: num  64 73 61 76 72 80 71 83 83 89 ...
 $ V2: num  4 4 4 4 6 6 6 6 8 8 ...
 $ V3: num  2 4 2 4 2 4 2 4 2 4 ...
}}
\usage{
CH06PR05
}
\description{
CH06PR05
}
\keyword{datasets}

