% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH06PR09}
\alias{CH06PR09}
\title{CH06PR09}
\format{\preformatted{'data.frame':	52 obs. of  4 variables:
 $ V1: int  4264 4496 4317 4292 4945 4325 4110 4111 4161 4560 ...
 $ V2: int  305657 328476 317164 366745 265518 301995 269334 267631 296350 277223 ...
 $ V3: num  7.17 6.2 4.61 7.02 8.61 6.88 7.23 6.27 6.49 6.37 ...
 $ V4: int  0 0 0 0 1 0 0 0 0 0 ...
}}
\usage{
CH06PR09
}
\description{
CH06PR09
}
\keyword{datasets}

