% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH06PR12}
\alias{CH06PR12}
\title{CH06PR12}
\format{\preformatted{'data.frame':	5 obs. of  3 variables:
 $ V1: int  302000 245000 280000 350000 295000
 $ V2: num  7.2 7.4 6.9 7 6.7
 $ V3: int  0 0 0 0 1
}}
\usage{
CH06PR12
}
\description{
CH06PR12
}
\keyword{datasets}

