% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH06PR13}
\alias{CH06PR13}
\title{CH06PR13}
\format{\preformatted{'data.frame':	4 obs. of  3 variables:
 $ V1: int  230000 250000 280000 340000
 $ V2: num  7.5 7.3 7.1 6.9
 $ V3: int  0 0 0 0
}}
\usage{
CH06PR13
}
\description{
CH06PR13
}
\keyword{datasets}

