% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH06PR15}
\alias{CH06PR15}
\title{CH06PR15}
\format{\preformatted{'data.frame':	46 obs. of  4 variables:
 $ V1: int  48 57 66 70 89 36 46 54 26 77 ...
 $ V2: int  50 36 40 41 28 49 42 45 52 29 ...
 $ V3: int  51 46 48 44 43 54 50 48 62 50 ...
 $ V4: num  2.3 2.3 2.2 1.8 1.8 2.9 2.2 2.4 2.9 2.1 ...
}}
\usage{
CH06PR15
}
\description{
CH06PR15
}
\keyword{datasets}

