% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH06PR18}
\alias{CH06PR18}
\title{CH06PR18}
\format{\preformatted{'data.frame':	81 obs. of  5 variables:
 $ V1: num  13.5 12 10.5 15 14 10.5 14 16.5 17.5 16.5 ...
 $ V2: int  1 14 16 4 11 15 2 1 1 8 ...
 $ V3: num  5.02 8.19 3 10.7 8.97 ...
 $ V4: num  0.14 0.27 0 0.05 0.07 0.24 0.19 0.6 0 0.03 ...
 $ V5: int  123000 104079 39998 57112 60000 101385 31300 248172 215000 251015 ...
}}
\usage{
CH06PR18
}
\description{
CH06PR18
}
\keyword{datasets}

