% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH06PR20}
\alias{CH06PR20}
\title{CH06PR20}
\format{\preformatted{'data.frame':	4 obs. of  4 variables:
 $ V1: int  5 6 14 12
 $ V2: num  8.25 8.5 11.5 10.25
 $ V3: num  0 0.23 0.11 0
 $ V4: int  250000 270000 300000 310000
}}
\usage{
CH06PR20
}
\description{
CH06PR20
}
\keyword{datasets}

