% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH06PR21}
\alias{CH06PR21}
\title{CH06PR21}
\format{\preformatted{'data.frame':	3 obs. of  4 variables:
 $ V1: int  4 6 12
 $ V2: num  10 11.5 12.5
 $ V3: num  0.1 0 0.32
 $ V4: int  80000 120000 340000
}}
\usage{
CH06PR21
}
\description{
CH06PR21
}
\keyword{datasets}

