% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH06PR27}
\alias{CH06PR27}
\title{CH06PR27}
\format{\preformatted{'data.frame':	6 obs. of  3 variables:
 $ V1: num  42 33 75 28 91 55
 $ V2: num  7 4 16 3 21 8
 $ V3: num  33 41 7 49 5 31
}}
\usage{
CH06PR27
}
\description{
CH06PR27
}
\keyword{datasets}

