% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH07TA01}
\alias{CH07TA01}
\title{CH07TA01}
\format{\preformatted{'data.frame':	20 obs. of  4 variables:
 $ V1: num  19.5 24.7 30.7 29.8 19.1 25.6 31.4 27.9 22.1 25.5 ...
 $ V2: num  43.1 49.8 51.9 54.3 42.2 53.9 58.5 52.1 49.9 53.5 ...
 $ V3: num  29.1 28.2 37 31.1 30.9 23.7 27.6 30.6 23.2 24.8 ...
 $ V4: num  11.9 22.8 18.7 20.1 12.9 21.7 27.1 25.4 21.3 19.3 ...
}}
\usage{
CH07TA01
}
\description{
CH07TA01
}
\keyword{datasets}

