% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH07TA05}
\alias{CH07TA05}
\title{CH07TA05}
\format{\preformatted{'data.frame':	21 obs. of  3 variables:
 $ V1: num  174 164 244 155 182 ...
 $ V2: num  68.5 45.2 91.3 47.8 46.9 66.1 49.5 52 48.9 38.4 ...
 $ V3: num  16.7 16.8 18.2 16.3 17.3 18.2 15.9 17.2 16.6 16 ...
}}
\usage{
CH07TA05
}
\description{
CH07TA05
}
\keyword{datasets}

