% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH07TA06}
\alias{CH07TA06}
\title{CH07TA06}
\format{\preformatted{'data.frame':	8 obs. of  3 variables:
 $ V1: int  4 4 4 4 6 6 6 6
 $ V2: int  2 2 3 3 2 2 3 3
 $ V3: int  42 39 48 51 49 53 61 60
}}
\usage{
CH07TA06
}
\description{
CH07TA06
}
\keyword{datasets}

