% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH07TA08}
\alias{CH07TA08}
\title{CH07TA08}
\format{\preformatted{'data.frame':	4 obs. of  3 variables:
 $ V1: int  2 8 6 10
 $ V2: int  6 9 8 10
 $ V3: int  23 83 63 103
}}
\usage{
CH07TA08
}
\description{
CH07TA08
}
\keyword{datasets}

