% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH08PR06}
\alias{CH08PR06}
\title{CH08PR06}
\format{\preformatted{'data.frame':	27 obs. of  2 variables:
 $ V1: num  27.1 22.1 21.9 10.7 1.4 18.8 14.7 5.7 18.6 20.4 ...
 $ V2: num  23 19 25 12 8 12 11 8 17 18 ...
}}
\usage{
CH08PR06
}
\description{
CH08PR06
}
\keyword{datasets}

