% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH08PR15}
\alias{CH08PR15}
\title{CH08PR15}
\format{\preformatted{'data.frame':	45 obs. of  1 variable:
 $ V1: int  1 0 0 0 0 0 1 1 1 1 ...
}}
\usage{
CH08PR15
}
\description{
CH08PR15
}
\keyword{datasets}

