% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH08PR16}
\alias{CH08PR16}
\title{CH08PR16}
\format{\preformatted{'data.frame':	120 obs. of  1 variable:
 $ V1: int  0 1 0 1 0 1 1 1 1 0 ...
}}
\usage{
CH08PR16
}
\description{
CH08PR16
}
\keyword{datasets}

