% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH08PR24}
\alias{CH08PR24}
\title{CH08PR24}
\format{\preformatted{'data.frame':	64 obs. of  3 variables:
 $ V1: num  78.8 73.8 64.6 76.2 87.2 70.6 86 83.1 94.5 71.2 ...
 $ V2: num  76.4 74.3 69.6 73.6 76.8 72.7 79.2 75.6 78.1 76.9 ...
 $ V3: int  0 0 0 0 0 1 0 0 0 1 ...
}}
\usage{
CH08PR24
}
\description{
CH08PR24
}
\keyword{datasets}

