% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH08PR29}
\alias{CH08PR29}
\title{CH08PR29}
\format{\preformatted{'data.frame':	8 obs. of  2 variables:
 $ V1: num  1 1.5 1.1 1.3 1.9 0.8 1.2 1.4
 $ V2: int  12 1 123 17 415 71 283 38
}}
\usage{
CH08PR29
}
\description{
CH08PR29
}
\keyword{datasets}

