% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH08TA01}
\alias{CH08TA01}
\title{CH08TA01}
\format{\preformatted{'data.frame':	11 obs. of  3 variables:
 $ V1: int  150 86 49 288 157 131 184 109 279 235 ...
 $ V2: num  0.6 1 1.4 0.6 1 1 1 1.4 0.6 1 ...
 $ V3: int  10 10 10 20 20 20 20 20 30 30 ...
}}
\usage{
CH08TA01
}
\description{
CH08TA01
}
\keyword{datasets}

