% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH08TA02}
\alias{CH08TA02}
\title{CH08TA02}
\format{\preformatted{'data.frame':	20 obs. of  3 variables:
 $ V1: int  17 26 21 30 22 0 12 19 4 16 ...
 $ V2: int  151 92 175 31 104 277 210 120 290 238 ...
 $ V3: int  0 0 0 0 0 0 0 0 0 0 ...
}}
\usage{
CH08TA02
}
\description{
CH08TA02
}
\keyword{datasets}

