% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH08TA05}
\alias{CH08TA05}
\title{CH08TA05}
\format{\preformatted{'data.frame':	27 obs. of  3 variables:
 $ V1: int  218 248 360 351 470 394 332 321 410 260 ...
 $ V2: int  100 125 220 205 300 255 225 175 270 170 ...
 $ V3: int  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
CH08TA05
}
\description{
CH08TA05
}
\keyword{datasets}

