% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH09PR10}
\alias{CH09PR10}
\title{CH09PR10}
\format{\preformatted{'data.frame':	25 obs. of  5 variables:
 $ V1: num  88 80 96 76 80 73 58 116 104 99 ...
 $ V2: num  86 62 110 101 100 78 120 105 112 120 ...
 $ V3: num  110 97 107 117 101 85 77 122 119 89 ...
 $ V4: num  100 99 103 93 95 95 80 116 106 105 ...
 $ V5: num  87 100 103 95 88 84 74 102 105 97 ...
}}
\usage{
CH09PR10
}
\description{
CH09PR10
}
\keyword{datasets}

