% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH09PR13}
\alias{CH09PR13}
\title{CH09PR13}
\format{\preformatted{'data.frame':	19 obs. of  4 variables:
 $ V1: num  49 55 85 32 26 28 95 26 74 37 ...
 $ V2: num  45 30 11 30 39 42 17 63 25 32 ...
 $ V3: num  36 28 16 46 76 78 24 80 12 27 ...
 $ V4: num  45 40 42 40 43 27 36 42 52 35 ...
}}
\usage{
CH09PR13
}
\description{
CH09PR13
}
\keyword{datasets}

