% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH09PR15}
\alias{CH09PR15}
\title{CH09PR15}
\format{\preformatted{'data.frame':	33 obs. of  4 variables:
 $ V1: num  132 53 50 82 110 100 68 92 60 94 ...
 $ V2: num  0.71 1.48 2.21 1.43 0.68 0.76 1.12 0.92 1.55 0.94 ...
 $ V3: num  38 78 69 70 45 65 76 61 68 64 ...
 $ V4: num  71 69 85 100 59 73 63 81 74 87 ...
}}
\usage{
CH09PR15
}
\description{
CH09PR15
}
\keyword{datasets}

