% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH09PR22}
\alias{CH09PR22}
\title{CH09PR22}
\format{\preformatted{'data.frame':	25 obs. of  5 variables:
 $ V1: num  58 92 71 77 92 66 61 57 66 75 ...
 $ V2: num  65 85 93 95 102 63 81 111 67 91 ...
 $ V3: num  109 90 73 57 139 101 129 102 98 111 ...
 $ V4: num  88 104 91 95 101 93 88 83 98 96 ...
 $ V5: num  84 98 82 85 92 84 76 72 84 84 ...
}}
\usage{
CH09PR22
}
\description{
CH09PR22
}
\keyword{datasets}

