% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH09TA01}
\alias{CH09TA01}
\title{CH09TA01}
\format{\preformatted{'data.frame':	54 obs. of  10 variables:
 $ V1 : num  6.7 5.1 7.4 6.5 7.8 5.8 5.7 3.7 6 3.7 ...
 $ V2 : int  62 59 57 73 65 38 46 68 67 76 ...
 $ V3 : int  81 66 83 41 115 72 63 81 93 94 ...
 $ V4 : num  2.59 1.7 2.16 2.01 4.3 1.42 1.91 2.57 2.5 2.4 ...
 $ V5 : int  50 39 55 48 45 65 49 69 58 48 ...
 $ V6 : int  0 0 0 0 0 1 1 1 0 0 ...
 $ V7 : int  1 0 0 0 0 1 0 1 1 1 ...
 $ V8 : int  0 0 0 0 1 0 1 0 0 0 ...
 $ V9 : int  695 403 710 349 2343 348 518 749 1056 968 ...
 $ V10: num  6.54 6 6.57 5.85 7.76 ...
}}
\usage{
CH09TA01
}
\description{
CH09TA01
}
\keyword{datasets}

