% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH09TA05}
\alias{CH09TA05}
\title{CH09TA05}
\format{\preformatted{'data.frame':	54 obs. of  10 variables:
 $ V1 : num  7.1 4.9 6.4 5.7 6.1 8 6.8 4.7 7 6.7 ...
 $ V2 : int  23 66 90 35 42 27 34 63 47 69 ...
 $ V3 : int  78 91 35 70 69 83 51 36 67 65 ...
 $ V4 : num  1.93 3.05 1.06 2.13 2.25 2.03 1.27 1.71 1.6 2.91 ...
 $ V5 : int  45 34 39 68 70 35 44 36 54 61 ...
 $ V6 : int  0 1 1 1 0 1 0 0 0 1 ...
 $ V7 : int  1 0 0 0 0 1 0 0 1 0 ...
 $ V8 : int  0 0 1 0 1 0 0 1 0 1 ...
 $ V9 : int  302 767 487 242 705 716 266 361 460 1060 ...
 $ V10: num  5.71 6.64 6.19 5.49 6.56 ...
}}
\usage{
CH09TA05
}
\description{
CH09TA05
}
\keyword{datasets}

