% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH10PR13}
\alias{CH10PR13}
\title{CH10PR13}
\format{\preformatted{'data.frame':	44 obs. of  4 variables:
 $ V1: num  12.8 11.6 12.8 11.2 9 ...
 $ V2: num  5.6 4.1 3.7 4.8 3.4 6.1 7.7 4 2.8 3.2 ...
 $ V3: num  5.6 4.8 3.5 4.5 3.7 5.8 7.2 4 2.3 3 ...
 $ V4: num  3.8 4.8 3.6 5.2 2.9 3.4 3.8 3.8 2.9 2.8 ...
}}
\usage{
CH10PR13
}
\description{
CH10PR13
}
\keyword{datasets}

