% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH10TA01}
\alias{CH10TA01}
\title{CH10TA01}
\format{\preformatted{'data.frame':	18 obs. of  3 variables:
 $ V1: num  45 57.2 26.9 66.3 41 ...
 $ V2: int  6 4 5 7 5 10 1 8 6 4 ...
 $ V3: int  91 162 11 240 73 311 316 154 164 54 ...
}}
\usage{
CH10TA01
}
\description{
CH10TA01
}
\keyword{datasets}

