% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH10TA02}
\alias{CH10TA02}
\title{CH10TA02}
\format{\preformatted{'data.frame':	4 obs. of  3 variables:
 $ V1: int  301 327 246 187
 $ V2: int  14 19 12 11
 $ V3: int  25 32 22 15
}}
\usage{
CH10TA02
}
\description{
CH10TA02
}
\keyword{datasets}

