% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11FI12}
\alias{CH11FI12}
\title{CH11FI12}
\format{\preformatted{'data.frame':	705 obs. of  4 variables:
 $ V1: num  1.25 1.32 1.48 1.57 1.67 1.73 1.85 1.92 2.28 2.33 ...
 $ V2: int  19 23 28 23 28 20 23 21 29 21 ...
 $ V3: int  74 95 77 47 67 69 76 46 59 77 ...
 $ V4: Factor w/ 2 levels "Training","Validation": 1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
CH11FI12
}
\description{
CH11FI12
}
\keyword{datasets}

