% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11PR06}
\alias{CH11PR06}
\title{CH11PR06}
\format{\preformatted{'data.frame':	12 obs. of  2 variables:
 $ V1: num  77 70 85 50 62 70 55 63 88 57 ...
 $ V2: num  16 14 22 10 14 17 10 13 19 12 ...
}}
\usage{
CH11PR06
}
\description{
CH11PR06
}
\keyword{datasets}

