% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11PR07}
\alias{CH11PR07}
\title{CH11PR07}
\format{\preformatted{'data.frame':	12 obs. of  2 variables:
 $ V1: num  28 75 37 53 22 58 40 96 46 52 ...
 $ V2: num  200 400 300 400 200 300 300 400 200 400 ...
}}
\usage{
CH11PR07
}
\description{
CH11PR07
}
\keyword{datasets}

