% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11PR08}
\alias{CH11PR08}
\title{CH11PR08}
\format{\preformatted{'data.frame':	65 obs. of  4 variables:
 $ V1: num  58.8 34.8 163.7 70 55.5 ...
 $ V2: int  3 1 3 3 3 2 1 1 2 3 ...
 $ V3: num  4.49 2.92 29.54 9.92 0.14 ...
 $ V4: int  0 0 42 0 0 4 0 0 3 0 ...
}}
\usage{
CH11PR08
}
\description{
CH11PR08
}
\keyword{datasets}

