% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11PR09}
\alias{CH11PR09}
\title{CH11PR09}
\format{\preformatted{'data.frame':	8 obs. of  8 variables:
 $ V1: num  0 0.01 0.02 0.04 0.06 0.08 0.09 0.1
 $ V2: num  0.49 0.461 0.443 0.463 0.41 0.401 0.398 0.394
 $ V3: num  0.296 0.322 0.336 0.349 0.354 0.356 0.356 0.356
 $ V4: num  0.169 0.167 0.167 0.166 0.165 0.164 0.164 0.164
 $ V5: num  20.07 10.36 6.37 3.2 1.98 ...
 $ V6: num  20.72 10.67 6.55 3.27 2.07 ...
 $ V7: num  1.22 1.17 1.14 1.08 1.02 0.98 0.95 0.93
 $ V8: num  0.742 0.742 0.715 0.741 0.741 ...
}}
\usage{
CH11PR09
}
\description{
CH11PR09
}
\keyword{datasets}

