% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11PR10}
\alias{CH11PR10}
\title{CH11PR10}
\format{\preformatted{'data.frame':	20 obs. of  3 variables:
 $ V1: num  58 152 41 93 101 38 203 78 117 44 ...
 $ V2: num  7 18 5 14 11 5 23 9 16 5 ...
 $ V3: num  5.11 16.72 3.2 7.03 10.98 ...
}}
\usage{
CH11PR10
}
\description{
CH11PR10
}
\keyword{datasets}

