% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11PR10b}
\alias{CH11PR10b}
\title{CH11PR10b}
\format{\preformatted{'data.frame':	8 obs. of  5 variables:
 $ V1: num  0 0.005 0.01 0.05 0.07 0.09 0.1 0.2
 $ V2: num  0.451 0.453 0.455 0.46 0.46 0.459 0.458 0.444
 $ V3: num  0.561 0.556 0.552 0.526 0.517 0.508 0.504 0.473
 $ V4: num  7.03 6.2 5.51 2.65 2.03 1.61 1.46 0.71
 $ V5: num  0.987 0.987 0.987 0.986 0.986 ...
}}
\usage{
CH11PR10b
}
\description{
CH11PR10b
}
\keyword{datasets}

