% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11PR11}
\alias{CH11PR11}
\title{CH11PR11}
\format{\preformatted{'data.frame':	2 obs. of  2 variables:
 $ V1: int  132 166
 $ V2: int  6 5
}}
\usage{
CH11PR11
}
\description{
CH11PR11
}
\keyword{datasets}

