% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11PR12}
\alias{CH11PR12}
\title{CH11PR12}
\format{\preformatted{'data.frame':	20 obs. of  2 variables:
 $ V1: int  185 195 216 156 179 184 157 171 176 128 ...
 $ V2: int  74 65 72 68 69 75 68 69 70 64 ...
}}
\usage{
CH11PR12
}
\description{
CH11PR12
}
\keyword{datasets}

