% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11PR23}
\alias{CH11PR23}
\title{CH11PR23}
\format{\preformatted{'data.frame':	13 obs. of  5 variables:
 $ V1: num  78.5 74.3 104.3 87.6 95.9 ...
 $ V2: int  7 1 11 11 7 11 3 1 2 21 ...
 $ V3: int  26 29 56 31 52 55 71 31 54 47 ...
 $ V4: int  6 15 8 8 6 9 17 22 18 4 ...
 $ V5: int  60 52 20 47 33 22 6 44 22 26 ...
}}
\usage{
CH11PR23
}
\description{
CH11PR23
}
\keyword{datasets}

