% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11PR25}
\alias{CH11PR25}
\title{CH11PR25}
\format{\preformatted{'data.frame':	25 obs. of  3 variables:
 $ V1: num  49.2 48.1 48 49.6 47 51.5 51.7 50.4 51.2 48.4 ...
 $ V2: num  6 6 6 6 6 8 8 8 8 8 ...
 $ V3: num  20 21 22 23 24 20 21 22 23 24 ...
}}
\usage{
CH11PR25
}
\description{
CH11PR25
}
\keyword{datasets}

