% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11PR28}
\alias{CH11PR28}
\title{CH11PR28}
\format{\preformatted{'data.frame':	12 obs. of  2 variables:
 $ V1: num  22 20 28 31 37 38 41 39 34 37 ...
 $ V2: num  35 35 40 40 45 45 50 50 55 55 ...
}}
\usage{
CH11PR28
}
\description{
CH11PR28
}
\keyword{datasets}

