% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11TA04}
\alias{CH11TA04}
\title{CH11TA04}
\format{\preformatted{'data.frame':	40 obs. of  7 variables:
 $ V1: Factor w/ 40 levels "Alabama","Arizona",..: 1 2 3 4 5 6 8 7 9 10 ...
 $ V2: int  252 259 256 256 267 270 261 231 255 258 ...
 $ V3: int  75 75 77 78 78 79 75 47 75 73 ...
 $ V4: int  78 73 77 68 85 86 83 76 73 80 ...
 $ V5: int  34 41 28 42 38 43 32 24 31 36 ...
 $ V6: int  18 12 20 11 9 12 18 33 19 17 ...
 $ V7: int  18 26 23 28 25 22 28 37 27 22 ...
}}
\usage{
CH11TA04
}
\description{
CH11TA04
}
\keyword{datasets}

