% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11TA07}
\alias{CH11TA07}
\title{CH11TA07}
\format{\preformatted{'data.frame':	18 obs. of  3 variables:
 $ V1: num  66.3 41 73 45 57.2 ...
 $ V2: int  7 5 10 6 4 5 4 6 9 5 ...
 $ V3: int  240 73 311 136 183 13 35 61 319 30 ...
}}
\usage{
CH11TA07
}
\description{
CH11TA07
}
\keyword{datasets}

