% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11TA09}
\alias{CH11TA09}
\title{CH11TA09}
\format{\preformatted{'data.frame':	25 obs. of  2 variables:
 $ V1: int  80 30 50 90 70 60 120 80 100 50 ...
 $ V2: int  399 121 221 376 361 224 546 352 353 157 ...
}}
\usage{
CH11TA09
}
\description{
CH11TA09
}
\keyword{datasets}

