% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11TA10}
\alias{CH11TA10}
\title{CH11TA10}
\format{\preformatted{'data.frame':	54 obs. of  2 variables:
 $ V1: int  27 21 22 24 25 23 20 20 29 24 ...
 $ V2: int  73 66 63 75 71 70 65 70 79 72 ...
}}
\usage{
CH11TA10
}
\description{
CH11TA10
}
\keyword{datasets}

