% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH11TA11}
\alias{CH11TA11}
\title{CH11TA11}
\format{\preformatted{'data.frame':	121 obs. of  8 variables:
 $ V1: int  1616 1329 3933 3786 465 794 618 1150 1538 1769 ...
 $ V2: int  13404 52314 30982 25207 20594 11507 9379 24991 30982 24991 ...
 $ V3: int  2 2 2 2 2 2 2 2 2 2 ...
 $ V4: int  52 60 57 64 40 44 43 42 59 41 ...
 $ V5: int  2 2 2 2 2 2 2 2 2 2 ...
 $ V6: int  2 2 4 4 2 2 2 2 2 2 ...
 $ V7: int  5 5 5 5 5 5 5 5 5 5 ...
 $ V8: int  1 1 2 2 1 1 1 1 1 1 ...
}}
\usage{
CH11TA11
}
\description{
CH11TA11
}
\keyword{datasets}

