% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH12PR09}
\alias{CH12PR09}
\title{CH12PR09}
\format{\preformatted{'data.frame':	16 obs. of  2 variables:
 $ V1: num  102.9 101.5 100.8 98 97.3 ...
 $ V2: num  2.05 2.03 2 1.95 1.94 ...
}}
\usage{
CH12PR09
}
\description{
CH12PR09
}
\keyword{datasets}

