% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH12PR13}
\alias{CH12PR13}
\title{CH12PR13}
\format{\preformatted{'data.frame':	20 obs. of  2 variables:
 $ V1: num  220 204 207 222 211 ...
 $ V2: num  2.52 2.17 2.23 2.52 2.31 ...
}}
\usage{
CH12PR13
}
\description{
CH12PR13
}
\keyword{datasets}

