% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH12PR23}
\alias{CH12PR23}
\title{CH12PR23}
\format{\preformatted{'data.frame':	10 obs. of  2 variables:
 $ V1: num  0.764 0.509 -0.242 -1.808 -0.485 ...
 $ V2: num  100 115 120 90 85 75 70 95 105 110
}}
\usage{
CH12PR23
}
\description{
CH12PR23
}
\keyword{datasets}

