% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH12TA02}
\alias{CH12TA02}
\title{CH12TA02}
\format{\preformatted{'data.frame':	20 obs. of  2 variables:
 $ V1: num  21 21.4 22 21.5 22.4 ...
 $ V2: num  127 130 133 129 135 ...
}}
\usage{
CH12TA02
}
\description{
CH12TA02
}
\keyword{datasets}

