% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH13PR05}
\alias{CH13PR05}
\title{CH13PR05}
\format{\preformatted{'data.frame':	16 obs. of  2 variables:
 $ V1: num  0.65 0.46 0.34 0.26 0.17 0.15 0.06 0.04 0.63 0.5 ...
 $ V2: int  200 400 800 1200 1600 2000 3000 4000 200 400 ...
}}
\usage{
CH13PR05
}
\description{
CH13PR05
}
\keyword{datasets}

