% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH13PR09}
\alias{CH13PR09}
\title{CH13PR09}
\format{\preformatted{'data.frame':	20 obs. of  3 variables:
 $ V1: num  0.9 0.8 0.65 0.46 0.34 0.26 0.17 0.15 0.06 0.04 ...
 $ V2: num  1 2.5 5 10 20 30 40 50 75 100 ...
 $ V3: num  0 0 0 0 0 0 0 0 0 0 ...
}}
\usage{
CH13PR09
}
\description{
CH13PR09
}
\keyword{datasets}

