% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH13PR10}
\alias{CH13PR10}
\title{CH13PR10}
\format{\preformatted{'data.frame':	18 obs. of  2 variables:
 $ V1: num  2.1 2.5 4.9 5.5 7 8.4 9.6 10.2 11.4 12.5 ...
 $ V2: num  1 1.5 2 3 4 5 6 7.5 8.5 10 ...
}}
\usage{
CH13PR10
}
\description{
CH13PR10
}
\keyword{datasets}

