% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH13PR13}
\alias{CH13PR13}
\title{CH13PR13}
\format{\preformatted{'data.frame':	19 obs. of  2 variables:
 $ V1: num  0.5 2.3 3.4 11.5 10.9 24 25.3 39.6 37.9 54.7 ...
 $ V2: num  1 2 3 3.5 3.5 4 4 4.5 4.5 5 ...
}}
\usage{
CH13PR13
}
\description{
CH13PR13
}
\keyword{datasets}

