% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH13PR17}
\alias{CH13PR17}
\title{CH13PR17}
\format{\preformatted{'data.frame':	18 obs. of  3 variables:
 $ V1: num  12 32 103 20 61 198 38 133 406 8 ...
 $ V2: num  1 10 100 1 10 100 1 10 100 1 ...
 $ V3: num  1 1 1 10 10 10 100 100 100 1 ...
}}
\usage{
CH13PR17
}
\description{
CH13PR17
}
\keyword{datasets}

