% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH13TA01}
\alias{CH13TA01}
\title{CH13TA01}
\format{\preformatted{'data.frame':	15 obs. of  2 variables:
 $ V1: int  54 50 45 37 35 25 20 16 18 13 ...
 $ V2: int  2 5 7 10 14 19 26 31 34 38 ...
}}
\usage{
CH13TA01
}
\description{
CH13TA01
}
\keyword{datasets}

