% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH13TA04}
\alias{CH13TA04}
\title{CH13TA04}
\format{\preformatted{'data.frame':	30 obs. of  3 variables:
 $ V1: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V2: int  1 2 3 5 7 10 15 20 30 40 ...
 $ V3: num  0.483 0.539 0.618 0.707 0.762 0.815 0.881 0.919 0.964 0.959 ...
}}
\usage{
CH13TA04
}
\description{
CH13TA04
}
\keyword{datasets}

