% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH14PR07}
\alias{CH14PR07}
\title{CH14PR07}
\format{\preformatted{'data.frame':	30 obs. of  2 variables:
 $ V1: num  0 1 0 0 0 0 1 0 0 1 ...
 $ V2: num  30 30 30 31 32 33 34 35 35 35 ...
}}
\usage{
CH14PR07
}
\description{
CH14PR07
}
\keyword{datasets}

