% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH14PR09}
\alias{CH14PR09}
\title{CH14PR09}
\format{\preformatted{'data.frame':	27 obs. of  2 variables:
 $ V1: num  0 0 0 1 1 0 0 1 1 1 ...
 $ V2: num  474 432 453 481 619 584 399 582 638 624 ...
}}
\usage{
CH14PR09
}
\description{
CH14PR09
}
\keyword{datasets}

