% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH14PR11}
\alias{CH14PR11}
\title{CH14PR11}
\format{\preformatted{'data.frame':	6 obs. of  3 variables:
 $ V1: num  2 5 10 20 25 30
 $ V2: num  500 500 500 500 500 500
 $ V3: num  72 103 170 296 406 449
}}
\usage{
CH14PR11
}
\description{
CH14PR11
}
\keyword{datasets}

