% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH14PR12}
\alias{CH14PR12}
\title{CH14PR12}
\format{\preformatted{'data.frame':	6 obs. of  3 variables:
 $ V1: num  1 2 3 4 5 6
 $ V2: num  250 250 250 250 250 250
 $ V3: num  28 53 93 126 172 197
}}
\usage{
CH14PR12
}
\description{
CH14PR12
}
\keyword{datasets}

