% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH14PR13}
\alias{CH14PR13}
\title{CH14PR13}
\format{\preformatted{'data.frame':	33 obs. of  3 variables:
 $ V1: num  0 0 1 0 0 1 1 1 0 1 ...
 $ V2: num  32 45 60 53 25 68 82 38 67 92 ...
 $ V3: num  3 2 2 1 4 1 2 5 2 2 ...
}}
\usage{
CH14PR13
}
\description{
CH14PR13
}
\keyword{datasets}

