% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH14PR14}
\alias{CH14PR14}
\title{CH14PR14}
\format{\preformatted{'data.frame':	159 obs. of  4 variables:
 $ V1: int  0 0 1 0 0 0 0 0 0 0 ...
 $ V2: int  59 61 82 51 53 62 51 70 71 55 ...
 $ V3: int  52 55 51 70 70 49 69 54 65 58 ...
 $ V4: int  0 1 0 0 0 1 1 1 1 1 ...
}}
\usage{
CH14PR14
}
\description{
CH14PR14
}
\keyword{datasets}

