% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH14PR39}
\alias{CH14PR39}
\title{CH14PR39}
\format{\preformatted{'data.frame':	100 obs. of  5 variables:
 $ V1: int  1 1 2 0 2 1 0 0 2 2 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  0 0 1 1 0 1 1 0 1 1 ...
 $ V4: int  45 62 43 76 51 73 40 66 80 56 ...
 $ V5: int  70 66 64 48 72 39 54 37 81 60 ...
}}
\usage{
CH14PR39
}
\description{
CH14PR39
}
\keyword{datasets}

