% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH14TA01}
\alias{CH14TA01}
\title{CH14TA01}
\format{\preformatted{'data.frame':	25 obs. of  3 variables:
 $ V1: int  14 29 6 25 18 4 18 12 22 6 ...
 $ V2: int  0 0 0 1 1 0 0 0 1 0 ...
 $ V3: num  0.31 0.835 0.11 0.727 0.462 ...
}}
\usage{
CH14TA01
}
\description{
CH14TA01
}
\keyword{datasets}

