% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH14TA02}
\alias{CH14TA02}
\title{CH14TA02}
\format{\preformatted{'data.frame':	5 obs. of  4 variables:
 $ V1: int  5 10 15 20 30
 $ V2: int  200 200 200 200 200
 $ V3: int  30 55 70 100 137
 $ V4: num  0.15 0.275 0.35 0.5 0.685
}}
\usage{
CH14TA02
}
\description{
CH14TA02
}
\keyword{datasets}

