% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH14TA03}
\alias{CH14TA03}
\title{CH14TA03}
\format{\preformatted{'data.frame':	98 obs. of  6 variables:
 $ V1: int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2: int  33 35 6 60 18 26 6 31 26 37 ...
 $ V3: int  0 0 0 0 0 0 0 1 1 1 ...
 $ V4: int  0 0 0 0 1 1 1 0 0 0 ...
 $ V5: int  0 0 0 0 0 0 0 0 0 0 ...
 $ V6: int  0 0 0 0 1 0 0 1 1 0 ...
}}
\usage{
CH14TA03
}
\description{
CH14TA03
}
\keyword{datasets}

