% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH14TA13}
\alias{CH14TA13}
\title{CH14TA13}
\format{\preformatted{'data.frame':	102 obs. of  10 variables:
 $ V1 : int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2 : int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3 : int  1 1 1 1 1 1 1 1 1 1 ...
 $ V4 : int  0 0 0 0 0 0 0 0 0 0 ...
 $ V5 : int  0 0 0 0 0 0 0 0 0 0 ...
 $ V6 : int  150 124 128 128 133 130 154 110 104 133 ...
 $ V7 : int  0 1 0 1 0 0 0 1 0 0 ...
 $ V8 : int  0 0 0 0 0 0 0 0 0 0 ...
 $ V9 : int  0 0 0 0 1 1 0 0 1 0 ...
 $ V10: int  1 0 1 1 1 1 0 0 1 1 ...
}}
\usage{
CH14TA13
}
\description{
CH14TA13
}
\keyword{datasets}

