% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH14TA14}
\alias{CH14TA14}
\title{CH14TA14}
\format{\preformatted{'data.frame':	110 obs. of  6 variables:
 $ V1: int  9 6 28 11 4 4 0 14 16 13 ...
 $ V2: int  606 641 505 866 599 520 354 483 1034 456 ...
 $ V3: int  41393 23635 55475 64646 31972 41755 46014 34626 85207 33021 ...
 $ V4: int  3 18 27 31 7 23 26 1 13 32 ...
 $ V5: num  3.04 1.95 6.54 1.67 0.72 2.24 0.77 3.51 4.23 3.07 ...
 $ V6: num  6.32 8.89 2.05 5.81 8.11 6.81 9.27 7.92 4.4 6.03 ...
}}
\usage{
CH14TA14
}
\description{
CH14TA14
}
\keyword{datasets}

