% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH15FI01}
\alias{CH15FI01}
\title{CH15FI01}
\format{\preformatted{'data.frame':	110 obs. of  2 variables:
 $ V1: num  4.8 6.4 6.3 6 5.4 5.8 6.1 6.3 5 6.2 ...
 $ V2: Factor w/ 2 levels "Attended","NotAttend": 1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
CH15FI01
}
\description{
CH15FI01
}
\keyword{datasets}

