% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH15FI06}
\alias{CH15FI06}
\title{CH15FI06}
\format{\preformatted{'data.frame':	8 obs. of  2 variables:
 $ V1: int  910 810 1270 1220 1510 1290 1180 1030
 $ V2: Factor w/ 4 levels "High","Low","Medium",..: 2 2 3 3 1 1 4 4
}}
\usage{
CH15FI06
}
\description{
CH15FI06
}
\keyword{datasets}

