% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH15FI09}
\alias{CH15FI09}
\title{CH15FI09}
\format{\preformatted{'data.frame':	8 obs. of  3 variables:
 $ V1: int  910 810 1270 1220 1510 1290 1180 1030
 $ V2: Factor w/ 4 levels "High","Low","Medium",..: 2 2 3 3 1 1 4 4
 $ V3: Factor w/ 2 levels "Plant A","Plant B": 2 1 2 1 2 1 2 1
}}
\usage{
CH15FI09
}
\description{
CH15FI09
}
\keyword{datasets}

