% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH15TA01}
\alias{CH15TA01}
\title{CH15TA01}
\format{\preformatted{'data.frame':	20 obs. of  4 variables:
 $ V1: int  1 2 3 4 5 6 7 8 9 10 ...
 $ V2: num  0.59 0.69 0.82 0.8 0.66 0.67 0.76 0.8 0.72 0.86 ...
 $ V3: num  0.43 0.53 0.58 0.65 0.38 0.48 0.62 0.6 0.44 0.62 ...
 $ V4: num  -0.16 -0.16 -0.24 -0.15 -0.28 -0.19 -0.14 -0.2 -0.28 -0.24 ...
}}
\usage{
CH15TA01
}
\description{
CH15TA01
}
\keyword{datasets}

