% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH16PR07}
\alias{CH16PR07}
\title{CH16PR07}
\format{\preformatted{'data.frame':	27 obs. of  3 variables:
 $ V1: num  7.6 8.2 6.8 5.8 6.9 6.6 6.3 7.7 6 6.7 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 2 ...
 $ V3: int  1 2 3 4 5 6 7 8 9 1 ...
}}
\usage{
CH16PR07
}
\description{
CH16PR07
}
\keyword{datasets}

