% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH16PR08}
\alias{CH16PR08}
\title{CH16PR08}
\format{\preformatted{'data.frame':	15 obs. of  3 variables:
 $ V1: num  28 26 31 27 35 34 29 25 31 29 ...
 $ V2: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V3: int  1 2 3 4 5 1 2 3 4 5 ...
}}
\usage{
CH16PR08
}
\description{
CH16PR08
}
\keyword{datasets}

