% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH16PR09}
\alias{CH16PR09}
\title{CH16PR09}
\format{\preformatted{'data.frame':	24 obs. of  3 variables:
 $ V1: num  29 42 38 40 43 40 30 42 30 35 ...
 $ V2: int  1 1 1 1 1 1 1 1 2 2 ...
 $ V3: int  1 2 3 4 5 6 7 8 1 2 ...
}}
\usage{
CH16PR09
}
\description{
CH16PR09
}
\keyword{datasets}

