% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH16PR10}
\alias{CH16PR10}
\title{CH16PR10}
\format{\preformatted{'data.frame':	36 obs. of  3 variables:
 $ V1: num  23 25 21 22 21 22 20 23 19 22 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 2 3 4 5 6 7 8 9 10 ...
}}
\usage{
CH16PR10
}
\description{
CH16PR10
}
\keyword{datasets}

