% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH16PR11}
\alias{CH16PR11}
\title{CH16PR11}
\format{\preformatted{'data.frame':	120 obs. of  3 variables:
 $ V1: num  -0.14 0.2 0.07 0.18 0.38 0.1 -0.04 -0.27 0.27 -0.21 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 2 3 4 5 6 7 8 9 10 ...
}}
\usage{
CH16PR11
}
\description{
CH16PR11
}
\keyword{datasets}

