% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH16PR12}
\alias{CH16PR12}
\title{CH16PR12}
\format{\preformatted{'data.frame':	100 obs. of  3 variables:
 $ V1: num  24 24 29 20 21 25 28 27 23 21 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 2 3 4 5 6 7 8 9 10 ...
}}
\usage{
CH16PR12
}
\description{
CH16PR12
}
\keyword{datasets}

