% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH16TA01}
\alias{CH16TA01}
\title{CH16TA01}
\format{\preformatted{'data.frame':	19 obs. of  3 variables:
 $ V1: int  11 17 16 14 15 12 10 15 19 11 ...
 $ V2: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V3: int  1 2 3 4 5 1 2 3 4 5 ...
}}
\usage{
CH16TA01
}
\description{
CH16TA01
}
\keyword{datasets}

