% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH16TA04}
\alias{CH16TA04}
\title{CH16TA04}
\format{\preformatted{'data.frame':	19 obs. of  6 variables:
 $ V1: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V2: int  1 2 3 4 5 1 2 3 4 5 ...
 $ V3: int  11 17 16 14 15 12 10 15 19 11 ...
 $ V4: int  1 1 1 1 1 0 0 0 0 0 ...
 $ V5: int  0 0 0 0 0 1 1 1 1 1 ...
 $ V6: int  0 0 0 0 0 0 0 0 0 0 ...
}}
\usage{
CH16TA04
}
\description{
CH16TA04
}
\keyword{datasets}

