% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH17TA02}
\alias{CH17TA02}
\title{CH17TA02}
\format{\preformatted{'data.frame':	40 obs. of  3 variables:
 $ V1: num  43.9 39 46.7 43.8 44.2 47.7 43.6 38.9 43.6 40 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 2 3 4 5 6 7 8 9 10 ...
}}
\usage{
CH17TA02
}
\description{
CH17TA02
}
\keyword{datasets}

