% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH17TA04}
\alias{CH17TA04}
\title{CH17TA04}
\format{\preformatted{'data.frame':	28 obs. of  3 variables:
 $ V1: int  40 39 39 36 42 43 41 53 48 49 ...
 $ V2: int  1 1 1 1 1 1 1 2 2 2 ...
 $ V3: int  1 2 3 4 5 6 7 1 2 3 ...
}}
\usage{
CH17TA04
}
\description{
CH17TA04
}
\keyword{datasets}

