% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH17TA05}
\alias{CH17TA05}
\title{CH17TA05}
\format{\preformatted{'data.frame':	28 obs. of  5 variables:
 $ V1: int  1 1 1 1 1 1 1 2 2 2 ...
 $ V2: int  1 2 3 4 5 6 7 1 2 3 ...
 $ V3: int  40 39 39 36 42 43 41 53 48 49 ...
 $ V4: int  -3 -3 -3 -3 -3 -3 -3 -1 -1 -1 ...
 $ V5: int  9 9 9 9 9 9 9 1 1 1 ...
}}
\usage{
CH17TA05
}
\description{
CH17TA05
}
\keyword{datasets}

