% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH18PR10}
\alias{CH18PR10}
\title{CH18PR10}
\format{\preformatted{'data.frame':	80 obs. of  3 variables:
 $ V1: num  0.1 0.28 0.1 0.47 0.83 0.65 0.28 -0.08 0.1 -0.26 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 2 3 4 5 6 7 8 9 10 ...
}}
\usage{
CH18PR10
}
\description{
CH18PR10
}
\keyword{datasets}

