% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH18PR15}
\alias{CH18PR15}
\title{CH18PR15}
\format{\preformatted{'data.frame':	80 obs. of  3 variables:
 $ V1: num  4 3 5 4 6 3 2 5 7 1 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 2 3 4 5 6 7 8 9 10 ...
}}
\usage{
CH18PR15
}
\description{
CH18PR15
}
\keyword{datasets}

