% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH18PR17}
\alias{CH18PR17}
\title{CH18PR17}
\format{\preformatted{'data.frame':	64 obs. of  3 variables:
 $ V1: num  4 3 2 3 4 4 3 6 5 4 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 2 3 4 5 6 7 8 9 10 ...
}}
\usage{
CH18PR17
}
\description{
CH18PR17
}
\keyword{datasets}

