% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH18PR25}
\alias{CH18PR25}
\title{CH18PR25}
\format{\preformatted{'data.frame':	30 obs. of  3 variables:
 $ V1: num  666 920 495 602 1499 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 2 3 4 5 6 7 8 9 10 ...
}}
\usage{
CH18PR25
}
\description{
CH18PR25
}
\keyword{datasets}

