% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH18TA01}
\alias{CH18TA01}
\title{CH18TA01}
\format{\preformatted{'data.frame':	10 obs. of  4 variables:
 $ V1: num  0.76 -4.14 3.56 0.66 1.06 4.56 0.46 -4.24 0.46 -3.14
 $ V2: num  0.36 -2.34 3.26 1.16 -1.74 2.96 -3.34 -1.34 1.36 -0.34
 $ V3: num  0.45 1.35 0.55 -1.55 2.05 0.15 2.65 -2.75 -4.15 1.25
 $ V4: num  -4.27 4.73 0.23 0.03 -1.17 -0.17 2.73 -1.77 0.43 -0.77
}}
\usage{
CH18TA01
}
\description{
CH18TA01
}
\keyword{datasets}

