% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH18TA02}
\alias{CH18TA02}
\title{CH18TA02}
\format{\preformatted{'data.frame':	40 obs. of  3 variables:
 $ V1: num  14.9 16.8 15.8 15.5 13.6 ...
 $ V2: int  1 1 1 1 1 1 1 1 2 2 ...
 $ V3: int  1 2 3 4 5 6 7 8 1 2 ...
}}
\usage{
CH18TA02
}
\description{
CH18TA02
}
\keyword{datasets}

