% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH18TA04}
\alias{CH18TA04}
\title{CH18TA04}
\format{\preformatted{'data.frame':	40 obs. of  9 variables:
 $ V1: int  1 1 1 1 1 1 1 1 2 2 ...
 $ V2: int  1 2 3 4 5 6 7 8 1 2 ...
 $ V3: num  14.9 16.8 15.8 15.5 13.6 ...
 $ V4: int  1 1 1 1 1 1 1 1 0 0 ...
 $ V5: int  0 0 0 0 0 0 0 0 1 1 ...
 $ V6: int  0 0 0 0 0 0 0 0 0 0 ...
 $ V7: int  0 0 0 0 0 0 0 0 0 0 ...
 $ V8: int  0 0 0 0 0 0 0 0 0 0 ...
 $ V9: num  0.653 0.653 0.653 0.653 0.653 0.653 0.653 0.653 0.637 0.637 ...
}}
\usage{
CH18TA04
}
\description{
CH18TA04
}
\keyword{datasets}

