% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH18TA05}
\alias{CH18TA05}
\title{CH18TA05}
\format{\preformatted{'data.frame':	15 obs. of  3 variables:
 $ V1: num  4.41 100.65 14.45 47.13 85.21 ...
 $ V2: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V3: int  1 2 3 4 5 1 2 3 4 5 ...
}}
\usage{
CH18TA05
}
\description{
CH18TA05
}
\keyword{datasets}

