% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH18TA07}
\alias{CH18TA07}
\title{CH18TA07}
\format{\preformatted{'data.frame':	39 obs. of  2 variables:
 $ V1: int  44 15 3 551 280 136 127 1024 65 1 ...
 $ V2: int  1 2 3 1 2 3 1 2 3 1 ...
}}
\usage{
CH18TA07
}
\description{
CH18TA07
}
\keyword{datasets}

