% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH19PR04}
\alias{CH19PR04}
\title{CH19PR04}
\format{\preformatted{'data.frame':	6 obs. of  3 variables:
 $ V1: int  34 23 36 40 29 42
 $ V2: int  1 1 1 2 2 2
 $ V3: int  1 2 3 1 2 3
}}
\usage{
CH19PR04
}
\description{
CH19PR04
}
\keyword{datasets}

