% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH19PR05}
\alias{CH19PR05}
\title{CH19PR05}
\format{\preformatted{'data.frame':	8 obs. of  3 variables:
 $ V1: int  250 265 268 269 288 273 270 269
 $ V2: int  1 1 1 1 2 2 2 2
 $ V3: int  1 2 3 4 1 2 3 4
}}
\usage{
CH19PR05
}
\description{
CH19PR05
}
\keyword{datasets}

