% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH19PR06}
\alias{CH19PR06}
\title{CH19PR06}
\format{\preformatted{'data.frame':	18 obs. of  4 variables:
 $ V1: int  80 80 80 90 90 90 75 80 90 80 ...
 $ V2: int  1 1 1 2 2 2 1 1 1 2 ...
 $ V3: int  1 2 3 1 2 3 1 2 3 1 ...
 $ V4: int  1 1 1 1 1 1 2 2 2 2 ...
}}
\usage{
CH19PR06
}
\description{
CH19PR06
}
\keyword{datasets}

