% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH19PR12}
\alias{CH19PR12}
\title{CH19PR12}
\format{\preformatted{'data.frame':	20 obs. of  4 variables:
 $ V1: num  11 7 12 6 10 15 12 14 11 16 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V4: int  1 2 3 4 5 1 2 3 4 5 ...
}}
\usage{
CH19PR12
}
\description{
CH19PR12
}
\keyword{datasets}

