% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH19PR14}
\alias{CH19PR14}
\title{CH19PR14}
\format{\preformatted{'data.frame':	36 obs. of  4 variables:
 $ V1: num  2.4 2.7 2.3 2.5 4.6 4.2 4.9 4.7 4.8 4.5 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V4: int  1 2 3 4 1 2 3 4 1 2 ...
}}
\usage{
CH19PR14
}
\description{
CH19PR14
}
\keyword{datasets}

