% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH19PR16}
\alias{CH19PR16}
\title{CH19PR16}
\format{\preformatted{'data.frame':	45 obs. of  4 variables:
 $ V1: num  62 48 63 57 69 57 45 39 54 44 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V4: int  1 2 3 4 5 1 2 3 4 5 ...
}}
\usage{
CH19PR16
}
\description{
CH19PR16
}
\keyword{datasets}

