% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH19PR18}
\alias{CH19PR18}
\title{CH19PR18}
\format{\preformatted{'data.frame':	60 obs. of  4 variables:
 $ V1: num  0 2 1 3 0 2 0 5 6 8 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V4: int  1 2 3 4 5 6 7 8 9 10 ...
}}
\usage{
CH19PR18
}
\description{
CH19PR18
}
\keyword{datasets}

