% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH19PR20}
\alias{CH19PR20}
\title{CH19PR20}
\format{\preformatted{'data.frame':	24 obs. of  4 variables:
 $ V1: num  240 206 217 225 110 118 103 95 56 60 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V4: int  1 2 3 4 1 2 3 4 1 2 ...
}}
\usage{
CH19PR20
}
\description{
CH19PR20
}
\keyword{datasets}

