% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH19TA07}
\alias{CH19TA07}
\title{CH19TA07}
\format{\preformatted{'data.frame':	12 obs. of  4 variables:
 $ V1: int  47 43 46 40 62 68 67 71 41 39 ...
 $ V2: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V3: int  1 1 2 2 1 1 2 2 1 1 ...
 $ V4: int  1 2 1 2 1 2 1 2 1 2 ...
}}
\usage{
CH19TA07
}
\description{
CH19TA07
}
\keyword{datasets}

