% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH20PR02}
\alias{CH20PR02}
\title{CH20PR02}
\format{\preformatted{'data.frame':	8 obs. of  3 variables:
 $ V1: num  16.5 21.4 11.8 17.3 12.3 16.9 16.6 21
 $ V2: int  1 1 2 2 3 3 4 4
 $ V3: int  1 2 1 2 1 2 1 2
}}
\usage{
CH20PR02
}
\description{
CH20PR02
}
\keyword{datasets}

