% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH20PR05}
\alias{CH20PR05}
\title{CH20PR05}
\format{\preformatted{'data.frame':	8 obs. of  3 variables:
 $ V1: num  18 22 31 32 15 23 29 33
 $ V2: int  1 1 1 1 2 2 2 2
 $ V3: int  1 2 3 4 1 2 3 4
}}
\usage{
CH20PR05
}
\description{
CH20PR05
}
\keyword{datasets}

