% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH20PR08}
\alias{CH20PR08}
\title{CH20PR08}
\format{\preformatted{'data.frame':	12 obs. of  3 variables:
 $ V1: num  13.9 14.2 20.5 24.8 15.7 16.3 21.7 23.6 15.1 15.4 ...
 $ V2: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V3: int  1 2 3 4 1 2 3 4 1 2 ...
}}
\usage{
CH20PR08
}
\description{
CH20PR08
}
\keyword{datasets}

