% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH20TA02}
\alias{CH20TA02}
\title{CH20TA02}
\format{\preformatted{'data.frame':	6 obs. of  3 variables:
 $ V1: int  140 100 210 180 220 200
 $ V2: int  1 1 2 2 3 3
 $ V3: int  1 2 1 2 1 2
}}
\usage{
CH20TA02
}
\description{
CH20TA02
}
\keyword{datasets}

