% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH21PR05}
\alias{CH21PR05}
\title{CH21PR05}
\format{\preformatted{'data.frame':	30 obs. of  3 variables:
 $ V1: num  73 81 92 76 78 89 75 76 87 74 ...
 $ V2: int  1 1 1 2 2 2 3 3 3 4 ...
 $ V3: int  1 2 3 1 2 3 1 2 3 1 ...
}}
\usage{
CH21PR05
}
\description{
CH21PR05
}
\keyword{datasets}

