% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH21PR07}
\alias{CH21PR07}
\title{CH21PR07}
\format{\preformatted{'data.frame':	15 obs. of  3 variables:
 $ V1: num  0.73 0.67 0.15 0.86 0.75 0.21 0.94 0.81 0.26 1.4 ...
 $ V2: int  1 1 1 2 2 2 3 3 3 4 ...
 $ V3: int  1 2 3 1 2 3 1 2 3 1 ...
}}
\usage{
CH21PR07
}
\description{
CH21PR07
}
\keyword{datasets}

