% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH21PR09}
\alias{CH21PR09}
\title{CH21PR09}
\format{\preformatted{'data.frame':	32 obs. of  4 variables:
 $ V1: num  0 0.6 0.5 1.2 0.3 0.7 0.6 1.3 0.4 0.8 ...
 $ V2: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V3: int  1 1 2 2 1 1 2 2 1 1 ...
 $ V4: int  1 2 1 2 1 2 1 2 1 2 ...
}}
\usage{
CH21PR09
}
\description{
CH21PR09
}
\keyword{datasets}

