% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH21PR13}
\alias{CH21PR13}
\title{CH21PR13}
\format{\preformatted{'data.frame':	30 obs. of  4 variables:
 $ V1: num  74 71 84 78 91 95 73 69 75 83 ...
 $ V2: int  1 1 1 1 1 1 2 2 2 2 ...
 $ V3: int  1 1 2 2 3 3 1 1 2 2 ...
 $ V4: int  1 2 1 2 1 2 1 2 1 2 ...
}}
\usage{
CH21PR13
}
\description{
CH21PR13
}
\keyword{datasets}

