% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH21TA01}
\alias{CH21TA01}
\title{CH21TA01}
\format{\preformatted{'data.frame':	15 obs. of  3 variables:
 $ V1: int  1 5 8 2 8 14 7 9 16 6 ...
 $ V2: int  1 1 1 2 2 2 3 3 3 4 ...
 $ V3: int  1 2 3 1 2 3 1 2 3 1 ...
}}
\usage{
CH21TA01
}
\description{
CH21TA01
}
\keyword{datasets}

