% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH21TA04}
\alias{CH21TA04}
\title{CH21TA04}
\format{\preformatted{'data.frame':	16 obs. of  4 variables:
 $ V1: int  12 8 7 5 3 9 5 9 14 16 ...
 $ V2: int  1 1 1 1 2 2 2 2 1 1 ...
 $ V3: int  1 1 1 1 1 1 1 1 2 2 ...
 $ V4: int  1 2 3 4 1 2 3 4 1 2 ...
}}
\usage{
CH21TA04
}
\description{
CH21TA04
}
\keyword{datasets}

