% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH22PR09}
\alias{CH22PR09}
\title{CH22PR09}
\format{\preformatted{'data.frame':	15 obs. of  4 variables:
 $ V1: num  28 26 31 27 35 34 29 25 31 29 ...
 $ V2: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V3: int  1 2 3 4 5 1 2 3 4 5 ...
 $ V4: num  300 381 226 350 100 153 334 473 264 325 ...
}}
\usage{
CH22PR09
}
\description{
CH22PR09
}
\keyword{datasets}

