% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH22PR11}
\alias{CH22PR11}
\title{CH22PR11}
\format{\preformatted{'data.frame':	24 obs. of  4 variables:
 $ V1: num  29 42 38 40 43 40 30 42 30 35 ...
 $ V2: int  1 1 1 1 1 1 1 1 2 2 ...
 $ V3: int  1 2 3 4 5 6 7 8 1 2 ...
 $ V4: num  18.3 30 26.5 28.1 29.7 27.8 19.8 29.3 20.8 25.2 ...
}}
\usage{
CH22PR11
}
\description{
CH22PR11
}
\keyword{datasets}

