% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH22PR13}
\alias{CH22PR13}
\title{CH22PR13}
\format{\preformatted{'data.frame':	15 obs. of  4 variables:
 $ V1: num  69 44 58 38 54 74 75 73 78 82 ...
 $ V2: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V3: int  1 2 3 4 5 1 2 3 4 5 ...
 $ V4: num  92 68 74 52 65 77 80 70 73 79 ...
}}
\usage{
CH22PR13
}
\description{
CH22PR13
}
\keyword{datasets}

