% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH22PR15}
\alias{CH22PR15}
\title{CH22PR15}
\format{\preformatted{'data.frame':	36 obs. of  5 variables:
 $ V1: num  21 23 19 22 22 23 21 22 20 21 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 1 1 1 2 2 2 2 ...
 $ V4: int  1 2 3 4 5 6 1 2 3 4 ...
 $ V5: num  3 5.1 1 4.4 2.7 4.9 3.5 4.2 2.2 3.1 ...
}}
\usage{
CH22PR15
}
\description{
CH22PR15
}
\keyword{datasets}

