% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH22PR17}
\alias{CH22PR17}
\title{CH22PR17}
\format{\preformatted{'data.frame':	20 obs. of  5 variables:
 $ V1: num  11 7 12 6 10 15 12 14 11 16 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V4: int  1 2 3 4 5 1 2 3 4 5 ...
 $ V5: num  42 30 47 31 35 51 35 48 38 49 ...
}}
\usage{
CH22PR17
}
\description{
CH22PR17
}
\keyword{datasets}

