% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH22PR19}
\alias{CH22PR19}
\title{CH22PR19}
\format{\preformatted{'data.frame':	30 obs. of  4 variables:
 $ V1: num  73 81 92 76 78 89 75 76 87 74 ...
 $ V2: int  1 1 1 2 2 2 3 3 3 4 ...
 $ V3: int  1 2 3 1 2 3 1 2 3 1 ...
 $ V4: num  93 98 91 94 93 94 89 91 92 86 ...
}}
\usage{
CH22PR19
}
\description{
CH22PR19
}
\keyword{datasets}

