% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH22TA01}
\alias{CH22TA01}
\title{CH22TA01}
\format{\preformatted{'data.frame':	15 obs. of  4 variables:
 $ V1: int  38 39 36 45 33 43 38 38 27 34 ...
 $ V2: int  21 26 22 28 19 34 26 29 18 25 ...
 $ V3: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V4: int  1 2 3 4 5 1 2 3 4 5 ...
}}
\usage{
CH22TA01
}
\description{
CH22TA01
}
\keyword{datasets}

