% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH22TA02}
\alias{CH22TA02}
\title{CH22TA02}
\format{\preformatted{'data.frame':	15 obs. of  9 variables:
 $ V1: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V2: int  1 2 3 4 5 1 2 3 4 5 ...
 $ V3: int  38 39 36 45 33 43 38 38 27 34 ...
 $ V4: int  21 26 22 28 19 34 26 29 18 25 ...
 $ V5: int  -4 1 -3 3 -6 9 1 4 -7 0 ...
 $ V6: int  1 1 1 1 1 0 0 0 0 0 ...
 $ V7: int  0 0 0 0 0 1 1 1 1 1 ...
 $ V8: int  -4 1 -3 3 -6 0 0 0 0 0 ...
 $ V9: int  0 0 0 0 0 9 1 4 -7 0 ...
}}
\usage{
CH22TA02
}
\description{
CH22TA02
}
\keyword{datasets}

