% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH22TA06}
\alias{CH22TA06}
\title{CH22TA06}
\format{\preformatted{'data.frame':	24 obs. of  5 variables:
 $ V1: int  98 60 77 80 95 64 55 60 75 65 ...
 $ V2: int  15 4 7 9 14 5 4 5 8 7 ...
 $ V3: int  1 1 1 1 1 1 2 2 2 2 ...
 $ V4: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V5: int  1 2 3 4 5 6 1 2 3 4 ...
}}
\usage{
CH22TA06
}
\description{
CH22TA06
}
\keyword{datasets}

