% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH23PR09}
\alias{CH23PR09}
\title{CH23PR09}
\format{\preformatted{'data.frame':	45 obs. of  4 variables:
 $ V1: num  1.7 1.9 1.8 2.1 2.5 2.7 2.9 2.5 2.6 2.8 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 2 2 3 3 3 3 3 3 ...
 $ V4: int  1 2 1 2 1 2 3 4 5 6 ...
}}
\usage{
CH23PR09
}
\description{
CH23PR09
}
\keyword{datasets}

