% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH23TA01}
\alias{CH23TA01}
\title{CH23TA01}
\format{\preformatted{'data.frame':	14 obs. of  4 variables:
 $ V1: num  1.4 2.4 2.2 2.1 1.7 0.7 1.1 2.4 2.5 1.8 ...
 $ V2: int  1 1 1 1 1 1 1 2 2 2 ...
 $ V3: int  1 1 1 2 2 3 3 1 2 2 ...
 $ V4: int  1 2 3 1 2 1 2 1 1 2 ...
}}
\usage{
CH23TA01
}
\description{
CH23TA01
}
\keyword{datasets}

