% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH23TA02}
\alias{CH23TA02}
\title{CH23TA02}
\format{\preformatted{'data.frame':	14 obs. of  9 variables:
 $ V1: int  1 1 1 1 1 1 1 2 2 2 ...
 $ V2: int  1 1 1 2 2 3 3 1 2 2 ...
 $ V3: int  1 2 3 1 2 1 2 1 1 2 ...
 $ V4: num  1.4 2.4 2.2 2.1 1.7 0.7 1.1 2.4 2.5 1.8 ...
 $ V5: int  1 1 1 1 1 1 1 -1 -1 -1 ...
 $ V6: int  1 1 1 0 0 -1 -1 1 0 0 ...
 $ V7: int  0 0 0 1 1 -1 -1 0 1 1 ...
 $ V8: int  1 1 1 0 0 -1 -1 -1 0 0 ...
 $ V9: int  0 0 0 1 1 -1 -1 0 -1 -1 ...
}}
\usage{
CH23TA02
}
\description{
CH23TA02
}
\keyword{datasets}

