% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH23TA06}
\alias{CH23TA06}
\title{CH23TA06}
\format{\preformatted{'data.frame':	8 obs. of  3 variables:
 $ V1: int  10 9 11 10 7 6 4 3
 $ V2: int  1 1 2 2 2 3 3 3
 $ V3: int  2 3 1 2 3 1 2 3
}}
\usage{
CH23TA06
}
\description{
CH23TA06
}
\keyword{datasets}

