% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH23TA08}
\alias{CH23TA08}
\title{CH23TA08}
\format{\preformatted{'data.frame':	14 obs. of  14 variables:
 $ V1 : int  1 1 1 1 1 1 1 2 2 2 ...
 $ V2 : int  1 1 1 2 2 3 3 1 2 2 ...
 $ V3 : int  1 2 3 1 2 1 2 1 1 2 ...
 $ V4 : num  1.4 2.4 2.2 2.1 1.7 0.7 1.1 2.4 2.5 1.8 ...
 $ V5 : int  1 1 1 0 0 0 0 0 0 0 ...
 $ V6 : int  0 0 0 1 1 0 0 0 0 0 ...
 $ V7 : int  0 0 0 0 0 1 1 0 0 0 ...
 $ V8 : int  0 0 0 0 0 0 0 1 0 0 ...
 $ V9 : int  0 0 0 0 0 0 0 0 1 1 ...
 $ V10: int  0 0 0 0 0 0 0 0 0 0 ...
 $ V11: int  1 1 1 0 0 0 0 -2 0 0 ...
 $ V12: int  0 0 0 1 1 0 0 2 0 0 ...
 $ V13: int  0 0 0 0 0 1 1 0 0 0 ...
 $ V14: int  0 0 0 0 0 0 0 1 1 1 ...
}}
\usage{
CH23TA08
}
\description{
CH23TA08
}
\keyword{datasets}

