% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH24PR04}
\alias{CH24PR04}
\title{CH24PR04}
\format{\preformatted{'data.frame':	12 obs. of  4 variables:
 $ V1: int  130 138 140 144 126 130 134 136 122 125 ...
 $ V2: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V3: int  1 2 1 2 1 2 1 2 1 2 ...
 $ V4: int  1 1 2 2 1 1 2 2 1 1 ...
}}
\usage{
CH24PR04
}
\description{
CH24PR04
}
\keyword{datasets}

