% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH24PR06}
\alias{CH24PR06}
\title{CH24PR06}
\format{\preformatted{'data.frame':	24 obs. of  5 variables:
 $ V1: num  39.9 32.2 36.3 56 56.9 56.6 53.5 50.7 52.8 70.9 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 1 1 1 2 2 2 2 ...
 $ V4: int  1 1 1 2 2 2 1 1 1 2 ...
 $ V5: int  1 2 3 1 2 3 1 2 3 1 ...
}}
\usage{
CH24PR06
}
\description{
CH24PR06
}
\keyword{datasets}

