% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH24PR09}
\alias{CH24PR09}
\title{CH24PR09}
\format{\preformatted{'data.frame':	48 obs. of  5 variables:
 $ V1: num  124 121 121 123 113 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 1 1 1 1 1 2 2 ...
 $ V4: int  1 1 1 1 2 2 2 2 1 1 ...
 $ V5: int  1 2 3 4 1 2 3 4 1 2 ...
}}
\usage{
CH24PR09
}
\description{
CH24PR09
}
\keyword{datasets}

