% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH24PR12}
\alias{CH24PR12}
\title{CH24PR12}
\format{\preformatted{'data.frame':	60 obs. of  5 variables:
 $ V1: num  1250 1175 1236 1239 1193 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V4: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V5: int  1 2 3 4 5 1 2 3 4 5 ...
}}
\usage{
CH24PR12
}
\description{
CH24PR12
}
\keyword{datasets}

