% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH24TA01}
\alias{CH24TA01}
\title{CH24TA01}
\format{\preformatted{'data.frame':	12 obs. of  4 variables:
 $ V1: int  9 12 18 19 20 21 9 10 14 19 ...
 $ V2: int  1 1 1 1 1 1 2 2 2 2 ...
 $ V3: int  1 2 3 1 2 3 1 2 3 1 ...
 $ V4: int  1 1 1 2 2 2 1 1 1 2 ...
}}
\usage{
CH24TA01
}
\description{
CH24TA01
}
\keyword{datasets}

