% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH24TA02}
\alias{CH24TA02}
\title{CH24TA02}
\format{\preformatted{'data.frame':	12 obs. of  4 variables:
 $ V1: num  10.5 12.5 16 17.5 19.5 23 9.5 10.5 13 18.5 ...
 $ V2: int  1 1 1 1 1 1 2 2 2 2 ...
 $ V3: int  1 2 3 1 2 3 1 2 3 1 ...
 $ V4: int  1 1 1 2 2 2 1 1 1 2 ...
}}
\usage{
CH24TA02
}
\description{
CH24TA02
}
\keyword{datasets}

