% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH24TA03}
\alias{CH24TA03}
\title{CH24TA03}
\format{\preformatted{'data.frame':	12 obs. of  4 variables:
 $ V1: num  10 12 15.5 18 20 23.5 10 11 13.5 18 ...
 $ V2: int  1 1 1 1 1 1 2 2 2 2 ...
 $ V3: int  1 2 3 1 2 3 1 2 3 1 ...
 $ V4: int  1 1 1 2 2 2 1 1 1 2 ...
}}
\usage{
CH24TA03
}
\description{
CH24TA03
}
\keyword{datasets}

