% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH24TA04}
\alias{CH24TA04}
\title{CH24TA04}
\format{\preformatted{'data.frame':	24 obs. of  5 variables:
 $ V1: num  24.1 29.2 24.6 20 21.9 17.6 14.6 15.3 12.3 16.1 ...
 $ V2: int  1 1 1 2 2 2 1 1 1 2 ...
 $ V3: int  1 1 1 1 1 1 2 2 2 2 ...
 $ V4: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V5: int  1 2 3 1 2 3 1 2 3 1 ...
}}
\usage{
CH24TA04
}
\description{
CH24TA04
}
\keyword{datasets}

