% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH24TA07}
\alias{CH24TA07}
\title{CH24TA07}
\format{\preformatted{'data.frame':	22 obs. of  12 variables:
 $ V1 : int  1 1 1 1 1 1 1 1 1 1 ...
 $ V2 : int  1 1 1 1 1 2 2 2 2 2 ...
 $ V3 : int  1 1 2 2 2 1 1 1 2 2 ...
 $ V4 : int  1 2 1 2 3 1 2 3 1 2 ...
 $ V5 : num  24.1 29.2 17.6 18.8 23.2 14.6 15.3 12.3 14.9 20.4 ...
 $ V6 : int  1 1 1 1 1 1 1 1 1 1 ...
 $ V7 : int  1 1 1 1 1 -1 -1 -1 -1 -1 ...
 $ V8 : int  1 1 -1 -1 -1 1 1 1 -1 -1 ...
 $ V9 : int  1 1 1 1 1 -1 -1 -1 -1 -1 ...
 $ V10: int  1 1 -1 -1 -1 1 1 1 -1 -1 ...
 $ V11: int  1 1 -1 -1 -1 -1 -1 -1 1 1 ...
 $ V12: int  1 1 -1 -1 -1 -1 -1 -1 1 1 ...
}}
\usage{
CH24TA07
}
\description{
CH24TA07
}
\keyword{datasets}

