% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH25PR07}
\alias{CH25PR07}
\title{CH25PR07}
\format{\preformatted{'data.frame':	48 obs. of  3 variables:
 $ V1: num  24.4 22.6 23.8 22 24.5 22.3 25 24.5 10.2 12.1 ...
 $ V2: int  1 1 1 1 1 1 1 1 2 2 ...
 $ V3: int  1 2 3 4 5 6 7 8 1 2 ...
}}
\usage{
CH25PR07
}
\description{
CH25PR07
}
\keyword{datasets}

