% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH25PR09}
\alias{CH25PR09}
\title{CH25PR09}
\format{\preformatted{'data.frame':	40 obs. of  3 variables:
 $ V1: num  205 204 207 202 208 206 209 205 207 206 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 2 3 4 5 6 7 8 9 10 ...
}}
\usage{
CH25PR09
}
\description{
CH25PR09
}
\keyword{datasets}

