% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH25PR15}
\alias{CH25PR15}
\title{CH25PR15}
\format{\preformatted{'data.frame':	40 obs. of  4 variables:
 $ V1: num  25.3 25.2 28.9 30 24.8 25.1 28.4 27.9 27.1 26.6 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 2 2 3 3 4 4 5 5 ...
 $ V4: int  1 2 1 2 1 2 1 2 1 2 ...
}}
\usage{
CH25PR15
}
\description{
CH25PR15
}
\keyword{datasets}

