% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH25PR17}
\alias{CH25PR17}
\title{CH25PR17}
\format{\preformatted{'data.frame':	48 obs. of  4 variables:
 $ V1: num  72 74.6 67.4 72.8 72.1 76.9 74.8 73.3 75.2 73.8 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V4: int  1 2 3 4 1 2 3 4 1 2 ...
}}
\usage{
CH25PR17
}
\description{
CH25PR17
}
\keyword{datasets}

