% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH25PR19}
\alias{CH25PR19}
\title{CH25PR19}
\format{\preformatted{'data.frame':	40 obs. of  3 variables:
 $ V1: num  11 13 10 18 15 20 28 15 30 18 ...
 $ V2: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V3: int  1 2 3 4 5 1 2 3 4 5 ...
}}
\usage{
CH25PR19
}
\description{
CH25PR19
}
\keyword{datasets}

