% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH25PR21}
\alias{CH25PR21}
\title{CH25PR21}
\format{\preformatted{'data.frame':	30 obs. of  3 variables:
 $ V1: int  10 15 14 8 12 13 21 27 25 14 ...
 $ V2: int  1 1 1 2 2 2 3 3 3 4 ...
 $ V3: int  1 2 3 1 2 3 1 2 3 1 ...
}}
\usage{
CH25PR21
}
\description{
CH25PR21
}
\keyword{datasets}

