% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH25TA01}
\alias{CH25TA01}
\title{CH25TA01}
\format{\preformatted{'data.frame':	20 obs. of  3 variables:
 $ V1: int  76 65 85 74 59 75 81 67 49 63 ...
 $ V2: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V3: int  1 2 3 4 1 2 3 4 1 2 ...
}}
\usage{
CH25TA01
}
\description{
CH25TA01
}
\keyword{datasets}

