% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH25TA12}
\alias{CH25TA12}
\title{CH25TA12}
\format{\preformatted{'data.frame':	39 obs. of  4 variables:
 $ V1: num  5.19 5.09 4.09 3.99 3.75 4.04 4.06 4.62 4.32 4.35 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 2 2 2 2 2 3 3 3 ...
 $ V4: int  1 2 1 2 3 4 5 1 2 3 ...
}}
\usage{
CH25TA12
}
\description{
CH25TA12
}
\keyword{datasets}

