% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH26PR04}
\alias{CH26PR04}
\title{CH26PR04}
\format{\preformatted{'data.frame':	60 obs. of  4 variables:
 $ V1: num  65 58 63 57 66 68 62 75 64 70 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V4: int  1 2 3 4 5 1 2 3 4 5 ...
}}
\usage{
CH26PR04
}
\description{
CH26PR04
}
\keyword{datasets}

