% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH26PR09}
\alias{CH26PR09}
\title{CH26PR09}
\format{\preformatted{'data.frame':	45 obs. of  4 variables:
 $ V1: num  42 56 35 40 28 26 38 42 35 53 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V4: int  1 2 3 4 5 1 2 3 4 5 ...
}}
\usage{
CH26PR09
}
\description{
CH26PR09
}
\keyword{datasets}

