% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH26PR14}
\alias{CH26PR14}
\title{CH26PR14}
\format{\preformatted{'data.frame':	14 obs. of  4 variables:
 $ V1: num  152 141 149 143 139 ...
 $ V2: int  1 1 1 1 1 1 1 2 2 2 ...
 $ V3: int  1 1 1 2 2 3 3 1 1 2 ...
 $ V4: int  1 2 3 1 2 1 2 1 2 1 ...
}}
\usage{
CH26PR14
}
\description{
CH26PR14
}
\keyword{datasets}

