% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH26PR17}
\alias{CH26PR17}
\title{CH26PR17}
\format{\preformatted{'data.frame':	30 obs. of  4 variables:
 $ V1: num  28 32 26 23 31 29 27 24 35 37 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 2 2 3 3 4 4 5 5 ...
 $ V4: int  1 2 1 2 1 2 1 2 1 2 ...
}}
\usage{
CH26PR17
}
\description{
CH26PR17
}
\keyword{datasets}

