% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH26PR19}
\alias{CH26PR19}
\title{CH26PR19}
\format{\preformatted{'data.frame':	36 obs. of  4 variables:
 $ V1: num  11.2 11.6 12 16.5 16.8 16.1 18.3 18.7 19 14.1 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 2 ...
 $ V3: int  1 1 1 2 2 2 3 3 3 1 ...
 $ V4: int  1 2 3 1 2 3 1 2 3 1 ...
}}
\usage{
CH26PR19
}
\description{
CH26PR19
}
\keyword{datasets}

