% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH26PR21}
\alias{CH26PR21}
\title{CH26PR21}
\format{\preformatted{'data.frame':	60 obs. of  4 variables:
 $ V1: num  2.3 2.1 2 2.5 2.3 2.5 2.6 2.4 2.7 2.4 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  1 1 1 2 2 2 3 3 3 4 ...
 $ V4: int  1 2 3 1 2 3 1 2 3 1 ...
}}
\usage{
CH26PR21
}
\description{
CH26PR21
}
\keyword{datasets}

