% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH26TA01}
\alias{CH26TA01}
\title{CH26TA01}
\format{\preformatted{'data.frame':	12 obs. of  4 variables:
 $ V1: int  25 29 14 11 11 6 22 18 17 20 ...
 $ V2: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V3: int  1 1 2 2 1 1 2 2 1 1 ...
 $ V4: int  1 2 1 2 1 2 1 2 1 2 ...
}}
\usage{
CH26TA01
}
\description{
CH26TA01
}
\keyword{datasets}

