% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH26TA06}
\alias{CH26TA06}
\title{CH26TA06}
\format{\preformatted{'data.frame':	9 obs. of  4 variables:
 $ V1: int  20 22 8 9 13 4 8 16 20
 $ V2: int  1 1 1 1 1 2 2 2 2
 $ V3: int  1 1 2 3 3 1 1 2 2
 $ V4: int  1 2 1 1 2 1 2 1 2
}}
\usage{
CH26TA06
}
\description{
CH26TA06
}
\keyword{datasets}

