% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH26TA09}
\alias{CH26TA09}
\title{CH26TA09}
\format{\preformatted{'data.frame':	18 obs. of  4 variables:
 $ V1: int  4 7 5 12 8 10 14 13 11 9 ...
 $ V2: int  1 1 1 1 1 1 2 2 2 2 ...
 $ V3: int  1 1 1 2 2 2 1 1 1 2 ...
 $ V4: int  1 2 3 1 2 3 1 2 3 1 ...
}}
\usage{
CH26TA09
}
\description{
CH26TA09
}
\keyword{datasets}

