% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH26TA12}
\alias{CH26TA12}
\title{CH26TA12}
\format{\preformatted{'data.frame':	16 obs. of  5 variables:
 $ V1: int  16 20 14 19 7 5 4 9 21 25 ...
 $ V2: int  1 1 1 1 2 2 2 2 1 1 ...
 $ V3: int  1 1 1 1 1 1 1 1 2 2 ...
 $ V4: int  1 1 2 2 1 1 2 2 1 1 ...
 $ V5: int  1 2 1 2 1 2 1 2 1 2 ...
}}
\usage{
CH26TA12
}
\description{
CH26TA12
}
\keyword{datasets}

