% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH27PR03}
\alias{CH27PR03}
\title{CH27PR03}
\format{\preformatted{'data.frame':	72 obs. of  3 variables:
 $ V1: num  21 21 23 35 36 48 19 24 27 36 ...
 $ V2: int  1 1 1 1 1 1 2 2 2 2 ...
 $ V3: int  1 2 3 4 5 6 1 2 3 4 ...
}}
\usage{
CH27PR03
}
\description{
CH27PR03
}
\keyword{datasets}

