% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH27PR06}
\alias{CH27PR06}
\title{CH27PR06}
\format{\preformatted{'data.frame':	24 obs. of  3 variables:
 $ V1: num  62.1 61.3 60.8 58.2 57.9 55.1 51.6 49.2 46.2 53.7 ...
 $ V2: int  1 1 1 2 2 2 3 3 3 4 ...
 $ V3: int  1 2 3 1 2 3 1 2 3 1 ...
}}
\usage{
CH27PR06
}
\description{
CH27PR06
}
\keyword{datasets}

