% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH27PR10}
\alias{CH27PR10}
\title{CH27PR10}
\format{\preformatted{'data.frame':	50 obs. of  3 variables:
 $ V1: num  3 1 2 5 4 4 2 1 3 5 ...
 $ V2: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V3: int  1 2 3 4 5 1 2 3 4 5 ...
}}
\usage{
CH27PR10
}
\description{
CH27PR10
}
\keyword{datasets}

