% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH27PR11}
\alias{CH27PR11}
\title{CH27PR11}
\format{\preformatted{'data.frame':	24 obs. of  4 variables:
 $ V1: num  10 18 16 35 14 19 19 32 17 18 ...
 $ V2: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V3: int  1 1 2 2 1 1 2 2 1 1 ...
 $ V4: int  1 2 1 2 1 2 1 2 1 2 ...
}}
\usage{
CH27PR11
}
\description{
CH27PR11
}
\keyword{datasets}

