% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH27PR13}
\alias{CH27PR13}
\title{CH27PR13}
\format{\preformatted{'data.frame':	32 obs. of  4 variables:
 $ V1: num  956 953 938 1049 769 ...
 $ V2: int  1 1 1 1 1 1 1 1 2 2 ...
 $ V3: int  1 1 1 1 2 2 2 2 1 1 ...
 $ V4: int  1 2 3 4 1 2 3 4 1 2 ...
}}
\usage{
CH27PR13
}
\description{
CH27PR13
}
\keyword{datasets}

