% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH27PR16}
\alias{CH27PR16}
\title{CH27PR16}
\format{\preformatted{'data.frame':	24 obs. of  4 variables:
 $ V1: num  3.1 7.5 2.5 5.1 3.8 8.1 2.8 5.3 3 7.6 ...
 $ V2: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V3: int  1 1 2 2 1 1 2 2 1 1 ...
 $ V4: int  1 2 1 2 1 2 1 2 1 2 ...
}}
\usage{
CH27PR16
}
\description{
CH27PR16
}
\keyword{datasets}

