% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH27PR18}
\alias{CH27PR18}
\title{CH27PR18}
\format{\preformatted{'data.frame':	40 obs. of  4 variables:
 $ V1: num  1.6 3.4 2.7 4.3 2.3 5.1 4.2 6.5 4.2 5.3 ...
 $ V2: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V3: int  1 1 2 2 1 1 2 2 1 1 ...
 $ V4: int  1 2 1 2 1 2 1 2 1 2 ...
}}
\usage{
CH27PR18
}
\description{
CH27PR18
}
\keyword{datasets}

