% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH27PR20}
\alias{CH27PR20}
\title{CH27PR20}
\format{\preformatted{'data.frame':	20 obs. of  4 variables:
 $ V1: num  43 48 63 70 40 43 52 53 31 36 ...
 $ V2: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V3: int  1 1 2 2 1 1 2 2 1 1 ...
 $ V4: int  1 2 1 2 1 2 1 2 1 2 ...
}}
\usage{
CH27PR20
}
\description{
CH27PR20
}
\keyword{datasets}

