% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH27TA02}
\alias{CH27TA02}
\title{CH27TA02}
\format{\preformatted{'data.frame':	24 obs. of  3 variables:
 $ V1: int  20 24 28 28 15 18 23 24 18 19 ...
 $ V2: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V3: int  1 2 3 4 1 2 3 4 1 2 ...
}}
\usage{
CH27TA02
}
\description{
CH27TA02
}
\keyword{datasets}

