% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH27TA04}
\alias{CH27TA04}
\title{CH27TA04}
\format{\preformatted{'data.frame':	30 obs. of  3 variables:
 $ V1: int  5 1 2 4 3 4 2 1 5 3 ...
 $ V2: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V3: int  1 2 3 4 5 1 2 3 4 5 ...
}}
\usage{
CH27TA04
}
\description{
CH27TA04
}
\keyword{datasets}

