% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH27TA07}
\alias{CH27TA07}
\title{CH27TA07}
\format{\preformatted{'data.frame':	30 obs. of  4 variables:
 $ V1: int  958 1005 351 549 730 1047 1122 436 632 784 ...
 $ V2: int  1 2 3 4 5 1 2 3 4 5 ...
 $ V3: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V4: int  1 1 1 1 1 2 2 2 2 2 ...
}}
\usage{
CH27TA07
}
\description{
CH27TA07
}
\keyword{datasets}

