% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH27TA08}
\alias{CH27TA08}
\title{CH27TA08}
\format{\preformatted{'data.frame':	40 obs. of  4 variables:
 $ V1: int  4 3 5 5 4 1 3 6 3 8 ...
 $ V2: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V3: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V4: int  1 2 3 4 5 1 2 3 4 5 ...
}}
\usage{
CH27TA08
}
\description{
CH27TA08
}
\keyword{datasets}

