% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH27TA12}
\alias{CH27TA12}
\title{CH27TA12}
\format{\preformatted{'data.frame':	48 obs. of  4 variables:
 $ V1: int  2 -1 0 3 1 2 -2 4 -2 -2 ...
 $ V2: int  1 2 3 4 5 6 7 8 9 10 ...
 $ V3: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V4: int  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
CH27TA12
}
\description{
CH27TA12
}
\keyword{datasets}

