% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH28PR06}
\alias{CH28PR06}
\title{CH28PR06}
\format{\preformatted{'data.frame':	36 obs. of  4 variables:
 $ V1: int  13 20 7 6 29 17 15 23 31 31 ...
 $ V2: int  3 8 4 4 9 2 3 6 9 9 ...
 $ V3: int  1 1 1 2 2 2 3 3 3 4 ...
 $ V4: int  1 2 3 1 2 3 1 2 3 1 ...
}}
\usage{
CH28PR06
}
\description{
CH28PR06
}
\keyword{datasets}

