% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH28PR08}
\alias{CH28PR08}
\title{CH28PR08}
\format{\preformatted{'data.frame':	12 obs. of  3 variables:
 $ V1: int  238 238 279 196 213 308 254 334 367 312 ...
 $ V2: Factor w/ 4 levels "A","B","C","D": 1 2 3 1 2 4 1 3 4 2 ...
 $ V3: int  1 1 1 2 2 2 3 3 3 4 ...
}}
\usage{
CH28PR08
}
\description{
CH28PR08
}
\keyword{datasets}

