% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH28PR17}
\alias{CH28PR17}
\title{CH28PR17}
\format{\preformatted{'data.frame':	25 obs. of  4 variables:
 $ V1: num  21 8 17 9 16 5 10 3 12 15 ...
 $ V2: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V3: int  1 2 3 4 5 1 2 3 4 5 ...
 $ V4: int  4 1 3 2 5 1 5 2 3 4 ...
}}
\usage{
CH28PR17
}
\description{
CH28PR17
}
\keyword{datasets}

