% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH28PR22}
\alias{CH28PR22}
\title{CH28PR22}
\format{\preformatted{'data.frame':	16 obs. of  4 variables:
 $ V1: num  0.02 0.15 0.45 0.18 0.27 0.24 -0.01 0.58 0.11 0.35 ...
 $ V2: int  1 1 1 1 2 2 2 2 3 3 ...
 $ V3: int  1 2 3 4 1 2 3 4 1 2 ...
 $ V4: int  1 2 4 3 2 3 1 4 3 4 ...
}}
\usage{
CH28PR22
}
\description{
CH28PR22
}
\keyword{datasets}

