% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH28PR24d}
\alias{CH28PR24d}
\title{CH28PR24d}
\format{\preformatted{'data.frame':	15 obs. of  4 variables:
 $ V1: num  1.2 1.5 1 1.7 1.4 1.9 1.5 2.8 2.1 2.7 ...
 $ V2: int  1 1 1 1 2 2 2 3 3 3 ...
 $ V3: int  1 2 3 4 1 2 4 1 2 3 ...
 $ V4: int  2 3 1 4 1 4 3 3 2 4 ...
}}
\usage{
CH28PR24d
}
\description{
CH28PR24d
}
\keyword{datasets}

