% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH28PR25}
\alias{CH28PR25}
\title{CH28PR25}
\format{\preformatted{'data.frame':	23 obs. of  4 variables:
 $ V1: num  8 17 9 16 5 10 3 12 15 20 ...
 $ V2: int  1 1 1 1 2 2 2 2 2 3 ...
 $ V3: int  2 3 4 5 1 2 3 4 5 1 ...
 $ V4: int  1 3 2 5 1 5 2 3 4 3 ...
}}
\usage{
CH28PR25
}
\description{
CH28PR25
}
\keyword{datasets}

