% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH28PR26}
\alias{CH28PR26}
\title{CH28PR26}
\format{\preformatted{'data.frame':	32 obs. of  5 variables:
 $ V1: num  83 86 64 69 78 75 76 74 70 76 ...
 $ V2: int  1 1 1 1 1 1 1 1 2 2 ...
 $ V3: int  1 1 2 2 3 3 4 4 1 1 ...
 $ V4: int  4 4 1 1 3 3 2 2 2 2 ...
 $ V5: int  1 2 1 2 1 2 1 2 1 2 ...
}}
\usage{
CH28PR26
}
\description{
CH28PR26
}
\keyword{datasets}

