% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH28PR28}
\alias{CH28PR28}
\title{CH28PR28}
\format{\preformatted{'data.frame':	27 obs. of  5 variables:
 $ V1: num  40 35 31 18 25 22 30 37 28 10 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 2 ...
 $ V3: int  1 1 1 2 2 2 3 3 3 1 ...
 $ V4: int  3 3 3 1 1 1 2 2 2 2 ...
 $ V5: int  1 2 3 1 2 3 1 2 3 1 ...
}}
\usage{
CH28PR28
}
\description{
CH28PR28
}
\keyword{datasets}

