% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH28TA02}
\alias{CH28TA02}
\title{CH28TA02}
\format{\preformatted{'data.frame':	30 obs. of  16 variables:
 $ V1 : int  1 1 1 2 2 2 3 3 3 4 ...
 $ V2 : int  1 2 3 1 2 4 1 2 5 1 ...
 $ V3 : int  6 6 8 3 7 7 6 8 6 3 ...
 $ V4 : int  1 1 1 0 0 0 0 0 0 0 ...
 $ V5 : int  0 0 0 1 1 1 0 0 0 0 ...
 $ V6 : int  0 0 0 0 0 0 1 1 1 0 ...
 $ V7 : int  0 0 0 0 0 0 0 0 0 1 ...
 $ V8 : int  0 0 0 0 0 0 0 0 0 0 ...
 $ V9 : int  0 0 0 0 0 0 0 0 0 0 ...
 $ V10: int  0 0 0 0 0 0 0 0 0 0 ...
 $ V11: int  0 0 0 0 0 0 0 0 0 0 ...
 $ V12: int  0 0 0 0 0 0 0 0 0 0 ...
 $ V13: int  1 0 0 1 0 0 1 0 -1 1 ...
 $ V14: int  0 1 0 0 1 0 0 1 -1 0 ...
 $ V15: int  0 0 1 0 0 0 0 0 -1 0 ...
 $ V16: int  0 0 0 0 0 1 0 0 -1 0 ...
}}
\usage{
CH28TA02
}
\description{
CH28TA02
}
\keyword{datasets}

