% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH28TA04}
\alias{CH28TA04}
\title{CH28TA04}
\format{\preformatted{'data.frame':	25 obs. of  4 variables:
 $ V1: int  18 13 7 17 21 17 34 29 13 26 ...
 $ V2: int  1 2 3 4 5 1 2 3 4 5 ...
 $ V3: int  1 1 1 1 1 2 2 2 2 2 ...
 $ V4: int  4 3 1 5 2 3 2 4 1 5 ...
}}
\usage{
CH28TA04
}
\description{
CH28TA04
}
\keyword{datasets}

