% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH28TA08}
\alias{CH28TA08}
\title{CH28TA08}
\format{\preformatted{'data.frame':	18 obs. of  4 variables:
 $ V1: int  19 16 24 22 10 14 20 24 14 15 ...
 $ V2: int  1 1 2 2 3 3 1 1 2 2 ...
 $ V3: int  1 1 1 1 1 1 2 2 2 2 ...
 $ V4: int  2 2 3 3 1 1 1 1 2 2 ...
}}
\usage{
CH28TA08
}
\description{
CH28TA08
}
\keyword{datasets}

