% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH28TA11}
\alias{CH28TA11}
\title{CH28TA11}
\format{\preformatted{'data.frame':	18 obs. of  5 variables:
 $ V1: int  9 4 12 13 7 5 12 12 14 14 ...
 $ V2: int  1 1 2 2 3 3 1 1 2 2 ...
 $ V3: int  1 1 1 1 1 1 2 2 2 2 ...
 $ V4: int  2 2 1 1 3 3 3 3 2 2 ...
 $ V5: int  1 2 1 2 1 2 1 2 1 2 ...
}}
\usage{
CH28TA11
}
\description{
CH28TA11
}
\keyword{datasets}

