% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH29PR07}
\alias{CH29PR07}
\title{CH29PR07}
\format{\preformatted{'data.frame':	32 obs. of  6 variables:
 $ V1: num  8.69 7.71 9.03 10.82 8.04 ...
 $ V2: int  -1 1 -1 1 -1 1 -1 1 -1 1 ...
 $ V3: int  -1 -1 1 1 -1 -1 1 1 -1 -1 ...
 $ V4: int  -1 -1 -1 -1 1 1 1 1 -1 -1 ...
 $ V5: int  -1 -1 -1 -1 -1 -1 -1 -1 1 1 ...
 $ V6: int  -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 ...
}}
\usage{
CH29PR07
}
\description{
CH29PR07
}
\keyword{datasets}

