% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH29PR09}
\alias{CH29PR09}
\title{CH29PR09}
\format{\preformatted{'data.frame':	19 obs. of  5 variables:
 $ V1: num  3.88 3.17 4.07 3.22 4 3.65 3.93 3.33 4.45 3.7 ...
 $ V2: int  -1 1 -1 1 -1 1 -1 1 -1 1 ...
 $ V3: int  -1 -1 1 1 -1 -1 1 1 -1 -1 ...
 $ V4: int  -1 -1 -1 -1 1 1 1 1 -1 -1 ...
 $ V5: int  -1 -1 -1 -1 -1 -1 -1 -1 1 1 ...
}}
\usage{
CH29PR09
}
\description{
CH29PR09
}
\keyword{datasets}

