% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH29PR16}
\alias{CH29PR16}
\title{CH29PR16}
\format{\preformatted{'data.frame':	8 obs. of  5 variables:
 $ V1: int  -1 1 -1 1 -1 1 -1 1
 $ V2: int  -1 -1 1 1 -1 -1 1 1
 $ V3: int  -1 -1 -1 -1 1 1 1 1
 $ V4: int  -1 -1 1 1 1 1 -1 -1
 $ V5: int  -1 1 1 -1 1 -1 -1 1
}}
\usage{
CH29PR16
}
\description{
CH29PR16
}
\keyword{datasets}

