% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH29PR18}
\alias{CH29PR18}
\title{CH29PR18}
\format{\preformatted{'data.frame':	16 obs. of  8 variables:
 $ V1: num  10.82 10.59 8.19 10.01 10.78 ...
 $ V2: int  -1 1 -1 1 -1 1 -1 1 -1 1 ...
 $ V3: int  -1 -1 1 1 -1 -1 1 1 -1 -1 ...
 $ V4: int  -1 -1 -1 -1 1 1 1 1 -1 -1 ...
 $ V5: int  -1 -1 -1 -1 -1 -1 -1 -1 1 1 ...
 $ V6: int  -1 1 1 -1 1 -1 -1 1 -1 1 ...
 $ V7: int  -1 -1 1 1 1 1 -1 -1 1 1 ...
 $ V8: int  -1 1 1 -1 -1 1 1 -1 1 -1 ...
}}
\usage{
CH29PR18
}
\description{
CH29PR18
}
\keyword{datasets}

