% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH29PR20}
\alias{CH29PR20}
\title{CH29PR20}
\format{\preformatted{'data.frame':	16 obs. of  10 variables:
 $ V1 : num  101.2 92.9 129.9 54.1 26.2 ...
 $ V2 : int  1 -1 1 -1 -1 -1 1 1 -1 1 ...
 $ V3 : int  1 -1 1 -1 -1 1 1 -1 1 -1 ...
 $ V4 : int  1 -1 1 1 1 -1 -1 -1 1 1 ...
 $ V5 : int  1 -1 1 1 1 1 -1 1 -1 -1 ...
 $ V6 : int  -1 -1 1 1 -1 -1 1 1 -1 -1 ...
 $ V7 : int  1 -1 -1 -1 1 -1 1 1 1 1 ...
 $ V8 : int  -1 -1 -1 1 1 -1 1 1 1 -1 ...
 $ V9 : int  -1 -1 -1 -1 -1 1 -1 1 1 1 ...
 $ V10: int  1 1 -1 1 -1 -1 1 -1 1 -1 ...
}}
\usage{
CH29PR20
}
\description{
CH29PR20
}
\keyword{datasets}

