% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH29PR26}
\alias{CH29PR26}
\title{CH29PR26}
\format{\preformatted{'data.frame':	64 obs. of  8 variables:
 $ V1: int  43 61 60 55 53 62 74 70 62 61 ...
 $ V2: int  1 1 1 1 1 1 1 1 2 2 ...
 $ V3: int  -1 1 -1 1 1 -1 1 -1 -1 1 ...
 $ V4: int  -1 1 1 -1 -1 1 1 -1 1 -1 ...
 $ V5: int  -1 1 1 -1 1 -1 -1 1 1 -1 ...
 $ V6: int  -1 1 -1 1 -1 1 -1 1 -1 1 ...
 $ V7: int  -1 -1 1 1 -1 -1 1 1 -1 -1 ...
 $ V8: int  -1 -1 -1 -1 1 1 1 1 -1 -1 ...
}}
\usage{
CH29PR26
}
\description{
CH29PR26
}
\keyword{datasets}

