% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH29PR29}
\alias{CH29PR29}
\title{CH29PR29}
\format{\preformatted{'data.frame':	22 obs. of  7 variables:
 $ V1: int  124 144 144 69 101 136 117 45 118 126 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  -1 1 1 -1 -1 1 1 -1 0 0 ...
 $ V4: int  -1 1 -1 1 -1 1 -1 1 0 0 ...
 $ V5: int  -1 -1 1 1 -1 -1 1 1 0 0 ...
 $ V6: int  -1 -1 -1 -1 1 1 1 1 0 0 ...
 $ V7: int  1 1 1 1 -1 -1 -1 -1 0 0 ...
}}
\usage{
CH29PR29
}
\description{
CH29PR29
}
\keyword{datasets}

