% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH29PR31}
\alias{CH29PR31}
\title{CH29PR31}
\format{\preformatted{'data.frame':	16 obs. of  7 variables:
 $ V1: int  -1 1 -1 1 -1 1 -1 1 -1 1 ...
 $ V2: int  -1 -1 1 1 -1 -1 1 1 -1 -1 ...
 $ V3: int  -1 -1 -1 -1 1 1 1 1 -1 -1 ...
 $ V4: int  -1 -1 -1 -1 -1 -1 -1 -1 1 1 ...
 $ V5: num  3.88 3.17 4.07 3.22 4 3.65 3.93 3.33 4.45 3.7 ...
 $ V6: num  3.1 2.75 3.05 2.42 3.1 2 2.6 2.01 3.25 2.62 ...
 $ V7: num  5.3 4.9 5.15 4.4 5.85 4.75 5.12 4.52 6.2 5.3 ...
}}
\usage{
CH29PR31
}
\description{
CH29PR31
}
\keyword{datasets}

