% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH29PR32}
\alias{CH29PR32}
\title{CH29PR32}
\format{\preformatted{'data.frame':	8 obs. of  10 variables:
 $ V1 : int  -1 1 -1 1 -1 1 -1 1
 $ V2 : int  -1 -1 1 1 -1 -1 1 1
 $ V3 : int  -1 -1 -1 -1 1 1 1 1
 $ V4 : int  -1 1 1 -1 1 -1 -1 1
 $ V5 : num  7.56 7.56 7.84 7.69 7.5 7.59 7.78 8.15
 $ V6 : num  7.62 7.81 7.7 8.09 7.56 7.56 7.83 8.1
 $ V7 : num  7.44 7.69 7.88 8.06 7.5 7.75 7.81 7.88
 $ V8 : num  7.52 7.81 7.32 7.56 7.5 7.63 7.5 7.6
 $ V9 : num  7.48 7.5 7.44 7.69 7.56 7.75 7.25 7.88
 $ V10: num  7.25 7.59 7.2 7.2 7.5 7.56 7.12 7.25
}}
\usage{
CH29PR32
}
\description{
CH29PR32
}
\keyword{datasets}

