% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH29TA01}
\alias{CH29TA01}
\title{CH29TA01}
\format{\preformatted{'data.frame':	24 obs. of  9 variables:
 $ V1: num  24.1 29.2 24.6 20 21.9 17.6 14.6 15.3 12.3 16.1 ...
 $ V2: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3: int  -1 -1 -1 1 1 1 -1 -1 -1 1 ...
 $ V4: int  -1 -1 -1 -1 -1 -1 1 1 1 1 ...
 $ V5: int  -1 -1 -1 -1 -1 -1 1 1 1 1 ...
 $ V6: int  1 1 1 -1 -1 -1 -1 -1 -1 1 ...
 $ V7: int  1 1 1 -1 -1 -1 -1 -1 -1 1 ...
 $ V8: int  1 1 1 1 1 1 1 1 1 1 ...
 $ V9: int  -1 -1 -1 1 1 1 -1 -1 -1 1 ...
}}
\usage{
CH29TA01
}
\description{
CH29TA01
}
\keyword{datasets}

