% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH29TA02}
\alias{CH29TA02}
\title{CH29TA02}
\format{\preformatted{'data.frame':	16 obs. of  17 variables:
 $ V1 : num  5.55 4.47 5.19 5.32 10.54 ...
 $ V2 : int  1 1 1 1 1 1 1 1 1 1 ...
 $ V3 : int  -1 1 -1 1 -1 1 -1 1 -1 1 ...
 $ V4 : int  -1 -1 1 1 -1 -1 1 1 -1 -1 ...
 $ V5 : int  -1 -1 -1 -1 1 1 1 1 -1 -1 ...
 $ V6 : int  -1 -1 -1 -1 -1 -1 -1 -1 1 1 ...
 $ V7 : int  1 -1 -1 1 1 -1 -1 1 1 -1 ...
 $ V8 : int  1 -1 1 -1 -1 1 -1 1 1 -1 ...
 $ V9 : int  1 -1 1 -1 1 -1 1 -1 -1 1 ...
 $ V10: int  1 1 -1 -1 -1 -1 1 1 1 1 ...
 $ V11: int  1 1 -1 -1 1 1 -1 -1 -1 -1 ...
 $ V12: int  1 1 1 1 -1 -1 -1 -1 -1 -1 ...
 $ V13: int  -1 1 1 -1 1 -1 -1 1 -1 1 ...
 $ V14: int  -1 1 1 -1 -1 1 1 -1 1 -1 ...
 $ V15: int  -1 1 -1 1 1 -1 1 -1 1 -1 ...
 $ V16: int  -1 -1 1 1 1 1 -1 -1 1 1 ...
 $ V17: int  1 -1 -1 1 -1 1 1 -1 -1 1 ...
}}
\usage{
CH29TA02
}
\description{
CH29TA02
}
\keyword{datasets}

