% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH29TA07}
\alias{CH29TA07}
\title{CH29TA07}
\format{\preformatted{'data.frame':	20 obs. of  7 variables:
 $ V1: int  -1 1 -1 1 -1 1 -1 1 -1 1 ...
 $ V2: int  -1 -1 1 1 -1 -1 1 1 -1 -1 ...
 $ V3: int  -1 -1 -1 -1 1 1 1 1 -1 -1 ...
 $ V4: int  -1 -1 -1 -1 -1 -1 -1 -1 1 1 ...
 $ V5: int  -1 1 1 -1 1 -1 -1 1 -1 1 ...
 $ V6: int  -1 -1 1 1 1 1 -1 -1 1 1 ...
 $ V7: int  4 6 7 2 3 1 5 9 3 2 ...
}}
\usage{
CH29TA07
}
\description{
CH29TA07
}
\keyword{datasets}

