% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH29TA11}
\alias{CH29TA11}
\title{CH29TA11}
\format{\preformatted{'data.frame':	8 obs. of  9 variables:
 $ V1: int  -1 1 -1 1 -1 1 -1 1
 $ V2: int  -1 -1 1 1 -1 -1 1 1
 $ V3: int  -1 -1 -1 -1 1 1 1 1
 $ V4: int  42 50 61 48 65 40 70 48
 $ V5: int  65 52 70 51 74 59 72 49
 $ V6: int  70 55 78 55 74 63 77 56
 $ V7: int  73 63 79 60 77 66 84 63
 $ V8: num  197.7 32.7 70 27 27 ...
 $ V9: num  5.29 3.49 4.25 3.3 3.3 ...
}}
\usage{
CH29TA11
}
\description{
CH29TA11
}
\keyword{datasets}

