% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH29TA12}
\alias{CH29TA12}
\title{CH29TA12}
\format{\preformatted{'data.frame':	16 obs. of  11 variables:
 $ V1 : int  -1 1 -1 1 -1 1 -1 1 -1 1 ...
 $ V2 : int  -1 -1 1 1 -1 -1 1 1 -1 -1 ...
 $ V3 : int  -1 -1 -1 -1 1 1 1 1 -1 -1 ...
 $ V4 : int  -1 1 1 -1 1 -1 -1 1 -1 1 ...
 $ V5 : int  -1 -1 -1 -1 -1 -1 -1 -1 1 1 ...
 $ V6 : num  0.8 3.2 0 5.8 0 0 0 1 9.8 6.4 ...
 $ V7 : num  0.4 0 0 0 3 1.2 2.6 2.3 2.5 3 ...
 $ V8 : num  0 0 0 0 0 0 0 5.2 13.8 13 ...
 $ V9 : num  0 0 2.4 2.8 2.4 4 1.2 0 2 0 ...
 $ V10: num  -1.92 0.94 0.365 2.036 0.912 ...
 $ V11: num  1.245 0.195 1.245 0.195 1.245 ...
}}
\usage{
CH29TA12
}
\description{
CH29TA12
}
\keyword{datasets}

