% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH30FI11}
\alias{CH30FI11}
\title{CH30FI11}
\format{\preformatted{'data.frame':	17 obs. of  4 variables:
 $ V1: int  63 12 64 26 59 11 62 31 55 0 ...
 $ V2: int  -1 1 -1 1 -1 1 -1 1 -1 1 ...
 $ V3: int  -1 -1 1 1 -1 -1 1 1 0 0 ...
 $ V4: int  1 1 1 1 -1 -1 -1 -1 0 0 ...
}}
\usage{
CH30FI11
}
\description{
CH30FI11
}
\keyword{datasets}

