% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH30PR11}
\alias{CH30PR11}
\title{CH30PR11}
\format{\preformatted{'data.frame':	15 obs. of  4 variables:
 $ V1: num  1.8 2 2.1 2.8 1.85 1.75 1.75 2.45 1.95 2.35 ...
 $ V2: int  -1 1 -1 1 -1 1 -1 1 -1 1 ...
 $ V3: int  -1 -1 1 1 -1 -1 1 1 0 0 ...
 $ V4: int  -1 -1 -1 -1 1 1 1 1 0 0 ...
}}
\usage{
CH30PR11
}
\description{
CH30PR11
}
\keyword{datasets}

