% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH30PR13}
\alias{CH30PR13}
\title{CH30PR13}
\format{\preformatted{'data.frame':	12 obs. of  3 variables:
 $ V1: int  144 192 112 215 126 179 174 176 196 191 ...
 $ V2: num  -1 1 -1 1 -1.41 ...
 $ V3: num  -1 -1 1 1 0 ...
}}
\usage{
CH30PR13
}
\description{
CH30PR13
}
\keyword{datasets}

