% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH30PR20}
\alias{CH30PR20}
\title{CH30PR20}
\format{\preformatted{'data.frame':	9 obs. of  4 variables:
 $ V1: num  -1 -0.75 -0.5 -0.5 0 0.25 0 0.5 1
 $ V2: int  -1 -1 -1 0 0 0 1 1 1
 $ V3: num  -1 -0.5 -0.5 0 0 0 0.25 0 1
 $ V4: int  -1 -1 0 0 0 0 0 1 1
}}
\usage{
CH30PR20
}
\description{
CH30PR20
}
\keyword{datasets}

