% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH30TA03}
\alias{CH30TA03}
\title{CH30TA03}
\format{\preformatted{'data.frame':	28 obs. of  4 variables:
 $ V1: int  -1 1 -1 1 -1 1 -1 1 -1 1 ...
 $ V2: int  -1 -1 1 1 -1 -1 1 1 -1 -1 ...
 $ V3: int  -1 -1 -1 -1 1 1 1 1 -1 -1 ...
 $ V4: int  -1 -1 -1 -1 -1 -1 -1 -1 1 1 ...
}}
\usage{
CH30TA03
}
\description{
CH30TA03
}
\keyword{datasets}

