% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH30TA04}
\alias{CH30TA04}
\title{CH30TA04}
\format{\preformatted{'data.frame':	9 obs. of  4 variables:
 $ V1: num  -1 1 -1 1 -1.41 ...
 $ V2: num  -1 -1 1 1 0 ...
 $ V3: num  -0.707 0.707 -0.707 0.707 -1 1 0 0 0
 $ V4: num  -0.707 -0.707 0.707 0.707 0 0 -1 1 0
}}
\usage{
CH30TA04
}
\description{
CH30TA04
}
\keyword{datasets}

