% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Scheffe.R
\name{Scheffe}
\alias{Scheffe}
\title{Scheffe Multiple Comparison Procedure}
\usage{
Scheffe(formula, data, cont, conf.level = 0.9, MSE)
}
\arguments{
\item{formula}{an object of class formula; a symbolic description of the model to be fitted.}

\item{data}{a data frame}

\item{cont}{a contrast matrix}

\item{conf.level}{confidence level}

\item{MSE}{value of mean squared error}
}
\description{
Computes post-hoc adjustment for multiple comparisons
}

